### 2020-06-11 4a34394
Corrected diving heuristic
