+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Periodic CVRP scheduling option
+
 ## [v0.2.0] - 07/06/2020
 
 ### Added
