### 2020-06-13 6a1211c
small fix, duration does not require custom REF
