### 2020-06-16 db16caf
greedy initial solution #31 #33
