+++ vrpy/main.py
@@ -2,6 +2,7 @@ from networkx import DiGraph, shortest_path, NetworkXError, has_path
 import logging
 from time import time
 
+from vrpy.greedy import Greedy
 from vrpy.master_solve_pulp import MasterSolvePulp
 from vrpy.subproblem_lp import SubProblemLP
 from vrpy.subproblem_cspy import SubProblemCSPY
@@ -344,7 +345,7 @@ class VehicleRoutingProblem:
                         # self._get_time_remaining(),
                         self._time_limit,
                         # 30,
-                        # exact=False,
+                        exact=False,
                     )
                     if self._more_routes:
                         break
@@ -484,8 +485,21 @@ class VehicleRoutingProblem:
                 self.G, self.load_capacity, self.duration, self.num_stops
             )
             alg.run()
-            logger.info("Initial solution found with value %s" % alg.best_value)
             self._initial_routes = alg.best_routes
+            logger.info(
+                "Clarke & Wright solution found with value %s and %s vehicles"
+                % (alg.best_value, len(alg.best_routes))
+            )
+
+            # Run greedy algorithm if possible
+            if not self.duration:
+                alg = Greedy(self.G, self.load_capacity)
+                alg.run()
+                logger.info(
+                    "Greedy solution found with value %s and %s vehicles"
+                    % (alg.best_value, len(alg.best_routes))
+                )
+                self._initial_routes += alg.best_routes
 
         # If pickup and delivery, initial routes are Source-pickup-delivery-Sink
         elif self.pickup_delivery:
@@ -520,7 +534,10 @@ class VehicleRoutingProblem:
             G.graph["vehicle_type"] = 0
             self._routes.append(G)
             for v in r:
-                self._routes_with_node[v] = [G]
+                if v in self._routes_with_node:
+                    self._routes_with_node[v].append(G)
+                else:
+                    self._routes_with_node[v] = [G]
 
     def _get_num_stops_upper_bound(self, max_capacity):
         """
