### 2020-06-16 c6cf35b
deactivate initial greedy, needs some fixing
