### 2020-06-17 7537249
fix initial greedy heuristic #31 #33
