### 2020-06-18 #2325 @arfon
This might also be a good fit for @bstabler?
