### 2020-06-18 #2325 @kyleniemeyer
If @bstabler is already overloaded, then I could, but this does seem like a better fit for him.
