### 2020-06-19 #2325 @arfon
> A bit off topic, but would be happy to pick this up (and possibly ask for @bstabler 's help if stuck?) If that's ok with you, feel free to assign me as editor 😃

Thanks @kakiac!
