### 2020-06-19 208c1d3
Initial commit for column-based formulation
- Reorganised main
- Only added set covering constraint by column
- Add method in master to add column
