+++ vrpy/main.py
@@ -83,8 +83,29 @@ class VehicleRoutingProblem:
         self.num_vehicles = num_vehicles
         self.periodic = periodic
         self.mixed_fleet = mixed_fleet
+        # Parameters for solving
+        self.masterproblem = None
+        self.routes = []
+        self._solver = None
+        self._time_limit = None
+        self._pricing_strategy = None
+        self._exact = None
+        self._cspy = None
+        self._dive = False
+        self._start_time = None
+        self._greedy = None
+        # parameters for column generation stopping criteria
+        self._more_routes = True
+        self._iteration = 0
+        self._no_improvement = 0
+        self._lower_bound = []
+        # Parameters for initial solution
         self._initial_routes = []
         self._preassignments = []
+        # Parameters for final solution
+        self._best_value = None
+        self._best_routes = []
+        self._best_routes_as_graphs = []
 
         # Check if given inputs are consistent
         self._check_vrp()
@@ -92,22 +113,20 @@ class VehicleRoutingProblem:
         # Keep track of paths containing nodes
         self._routes = []
         self._routes_with_node = {}
-        for v in self.G.nodes():
-            if v not in ["Source", "Sink"]:
-                self._routes_with_node[v] = []
+        for node in self.G.nodes():
+            if node not in ["Source", "Sink"]:
+                self._routes_with_node[node] = []
 
-    def solve(
-        self,
-        initial_routes=None,
-        preassignments=None,
-        pricing_strategy="BestEdges1",
-        cspy=True,
-        exact=True,
-        time_limit=None,
-        solver="cbc",
-        dive=False,
-        greedy=False,
-    ):
+    def solve(self,
+              initial_routes=None,
+              preassignments=None,
+              pricing_strategy="BestEdges1",
+              cspy=True,
+              exact=True,
+              time_limit=None,
+              solver="cbc",
+              dive=False,
+              greedy=False):
         """Iteratively generates columns with negative reduced cost and solves as MIP.
 
         Args:
@@ -155,7 +174,7 @@ class VehicleRoutingProblem:
         Returns:
             float: Optimal solution of MIP based on generated columns
         """
-        # Solving attributes
+        # set solving attributes
         self._solver = solver
         self._time_limit = time_limit
         self._pricing_strategy = pricing_strategy
@@ -170,14 +189,14 @@ class VehicleRoutingProblem:
             self._initial_routes = initial_routes
 
         # If only one type of vehicle, some formatting is done
-        if self.mixed_fleet == False:
+        if not self.mixed_fleet:
             self._format()
 
         # Pre-processing
         self._pre_solve()
 
         # Initialization
-        self._initialize()
+        self._initialize(solver)
 
         # Column generation
         self._column_generation()
@@ -190,22 +209,8 @@ class VehicleRoutingProblem:
             # Solve as MIP
             # FIXME issue #23
             try:
-                masterproblem_mip = MasterSolvePulp(
-                    self.G,
-                    self._routes_with_node,
-                    self._routes,
-                    self.drop_penalty,
-                    self.num_vehicles,
-                    self.periodic,
-                    self._solver,
-                    self._time_limit,
-                    # self._get_time_remaining(),
-                    relax=False,
-                )
-                (
-                    self._best_value,
-                    self._best_routes_as_graphs,
-                ) = masterproblem_mip.solve()
+                self._best_value, self._best_routes_as_graphs = self.masterproblem.solve(
+                    relax=False)
             except Exception:
                 self._best_value, self._best_routes_as_graphs = (
                     self._lower_bound[-1],
@@ -224,26 +229,10 @@ class VehicleRoutingProblem:
                 self.best_routes,
                 self.best_routes_type,
                 self.num_vehicles,
[…]
