+++ vrpy/master_solve_pulp.py
@@ -1,6 +1,7 @@
-import pulp
 import logging
+
 from networkx import shortest_path
+import pulp
 
 from vrpy.masterproblem import MasterProblemBase
 
@@ -14,26 +15,42 @@ class MasterSolvePulp(MasterProblemBase):
     Inherits problem parameters from MasterProblemBase
     """
 
-    def solve(self):
+    def __init__(self, *args, solver):
+        super(MasterSolvePulp, self).__init__(*args)
+        # create problem
+        self.prob = pulp.LpProblem("MasterProblem", pulp.LpMinimize)
+        # objective
+        self.objective = pulp.LpConstraintVar("objective")
+        # variables
+        self.y = {}
+        # constrs
+        self.set_covering_constrs = {}
+        self.drop_penalty_constrs = {}
+        # Diving attributes
+        self._tabu_list = []
+
         self._formulate()
+        self._set_solver(solver)
+
+    def solve(self, relax):
         # self.prob.writeLP("master.lp")
-        self._solve()
+        self._solve(relax)
         logger.debug("master problem")
         logger.debug("Status: %s" % pulp.LpStatus[self.prob.status])
         logger.debug("Objective: %s" % pulp.value(self.prob.objective))
 
         if pulp.LpStatus[self.prob.status] != "Optimal":
             raise Exception("problem " + str(pulp.LpStatus[self.prob.status]))
-        if self.relax:
+        if relax:
             for r in self.routes:
                 if pulp.value(self.y[r.graph["name"]]) > 0.5:
                     logger.debug("route %s selected" % r.graph["name"])
             duals = self.get_duals()
             logger.debug("duals : %s" % duals)
-            return duals, pulp.value(self.prob.objective)
+            return duals, self.prob.objective.value()
 
         else:
-            return self._get_total_cost_and_routes()
+            return self._get_total_cost_and_routes(relax=False)
 
     def solve_and_dive(self, max_depth=3, max_discrepancy=1):
         """
@@ -43,13 +60,12 @@ class MasterSolvePulp(MasterProblemBase):
 
         .. _Sadykov et al. (2019): https://pubsonline.informs.org/doi/abs/10.1287/ijoc.2018.0822
         """
-        self._formulate()
-        self._solve()
+        self._solve(relax=True)
 
         depth = 0
         tabu_list = []
         stop_diving = True
-        relax = self.relax.model.deepcopy()
+        relax = self.prob.deepcopy()
         constrs = {}
         while depth <= max_depth and len(tabu_list) <= max_discrepancy:
             non_integer_vars = list(
@@ -97,14 +113,20 @@ class MasterSolvePulp(MasterProblemBase):
                     break
                 tabu_list.append(var_to_fix.name)
                 depth += 1
-                self.relax.model.extend(constrs)
+                self.prob.extend(constrs)
                 logger.info("fixed %s with previous value %s", var_to_fix.name,
                             value_previous)
             else:
                 break
         logger.debug("Ran diving with LDS and fixed %s vars", len(tabu_list))
         self._tabu_list.extend(tabu_list)
-        return self._get_total_cost_and_routes(), stop_diving
+        return self._get_total_cost_and_routes(relax=True), stop_diving
+
+    def update(self, new_route):
+        """Add new column.
+        """
+        self._add_single_route_selecting_variable(new_route)
+        self.prob.writeLP("./check.lp")
 
     def get_duals(self):
         """Gets the dual values of each constraint of the master problem.
@@ -114,12 +136,12 @@ class MasterSolvePulp(MasterProblemBase):
         """
         duals = {}
         # set covering duals
-        for v in self.G.nodes():
-            if (v not in ["Source", "Sink"] and
[…]
