### 2020-06-19 4713e8a
Remove drop variable from RHS for constraint init
