### 2020-06-22 82af25e
initial greedy heuristic #33
