### 2020-06-25 e6780b5
Revert "Finish transfer to column based formulation"
