+++ vrpy/main.py
@@ -54,6 +54,7 @@ class VehicleRoutingProblem:
             True if heterogeneous fleet.
             Defaluts to False.
     """
+
     def __init__(
         self,
         G,
@@ -79,7 +80,7 @@ class VehicleRoutingProblem:
         self.distribution_collection = distribution_collection
         self.drop_penalty = drop_penalty
         self.fixed_cost = fixed_cost
-        self.num_vehicles = num_vehicles if num_vehicles is not None else []
+        self.num_vehicles = num_vehicles
         self.periodic = periodic
         self.mixed_fleet = mixed_fleet
         # Parameters for solving
@@ -94,7 +95,7 @@ class VehicleRoutingProblem:
         self._start_time = None
         self._greedy = None
         # parameters for column generation stopping criteria
-        self._more_routes = None
+        self._more_routes = True
         self._iteration = 0
         self._no_improvement = 0
         self._lower_bound = []
@@ -173,9 +174,7 @@ class VehicleRoutingProblem:
         Returns:
             float: Optimal solution of MIP based on generated columns
         """
-
         # set solving attributes
-        self._more_routes = True
         self._solver = solver
         self._time_limit = time_limit
         self._pricing_strategy = pricing_strategy
@@ -219,7 +218,7 @@ class VehicleRoutingProblem:
                 )
         # Get dropped nodes
         if self.drop_penalty:
-            self._dropped_nodes = self.masterproblem.dropped_nodes
+            self._dropped_nodes = masterproblem_mip.dropped_nodes
         # Convert best routes into lists of nodes
         self._best_routes_as_node_lists()
         # Schedule routes over time span if Periodic CVRP
@@ -312,9 +311,9 @@ class VehicleRoutingProblem:
         for vehicle in range(self._vehicle_types):
 
             # Solve pricing problem with randomised greedy algorithm
-            if (self._greedy and not self.time_windows
-                    and not self.distribution_collection
-                    and not self.pickup_delivery):
+            if (self._greedy and not self.time_windows and
+                    not self.distribution_collection and
+                    not self.pickup_delivery):
                 subproblem = self._def_subproblem(duals, vehicle, greedy=True)
                 self.routes, self._more_routes = subproblem.solve(n_runs=20)
 
@@ -377,8 +376,6 @@ class VehicleRoutingProblem:
                     self._time_limit,
                     # exact=False,
                 )
-            if self._more_routes:
-                self.masterproblem.update(self.routes[-1])
 
         # Keep track of convergence rate and update stopping criteria parameters
         self._iteration += 1
@@ -388,6 +385,8 @@ class VehicleRoutingProblem:
             self._no_improvement = 0
         self._lower_bound.append(relaxed_cost)
         # Add column (new route) to the master problem
+        if self._more_routes:
+            self.masterproblem.update(self.routes[-1])
 
         return stop_diving if self._dive else False
 
@@ -468,14 +467,13 @@ class VehicleRoutingProblem:
             - with a round trip otherwise.
         """
         # Run Clarke & Wright if possible
-        if (not self.time_windows and not self.pickup_delivery
-                and not self.distribution_collection and not self.mixed_fleet
-                and not self.periodic):
+        if (not self.time_windows and not self.pickup_delivery and
+                not self.distribution_collection and not self.mixed_fleet and
+                not self.periodic):
             alg = ClarkeWright(self.G, self.load_capacity, self.duration,
                                self.num_stops)
             alg.run()
-            logger.info("Initial solution found with value %s" %
-                        alg.best_value)
+            logger.info("Initial solution found with value %s" % alg.best_value)
             self._initial_routes = alg.best_routes
 
[…]
