### 2020-06-25 1951e89
Remove old column generation method
