+++ vrpy/main.py
@@ -55,6 +55,7 @@ class VehicleRoutingProblem:
             True if heterogeneous fleet.
             Defaluts to False.
     """
+
     def __init__(
         self,
         G,
@@ -109,7 +110,6 @@ class VehicleRoutingProblem:
 
         # Check if given inputs are consistent
         self._check_vrp()
-
 
     def solve(
         self,
@@ -232,7 +232,6 @@ class VehicleRoutingProblem:
             schedule.solve(self._get_time_remaining())
             self._schedule = schedule.routes_per_day
 
-
     def _column_generation(self):
         while self._more_routes:
             # Generate good columns
@@ -291,24 +290,6 @@ class VehicleRoutingProblem:
                                              self._get_time_remaining(),
                                              solver=solver)
 
-    def _column_generation(self):
-        while self._more_routes:
-            # Generate good columns
-            stop = self._find_columns()
-            # Stop if time limit is passed
-            # FIXME issue #23
-            # if self._get_time_remaining() and self._get_time_remaining() <= 0.0:
-            if self._time_limit and (time() - self._start_time >
-                                     self._time_limit):
-                logger.info("time up !")
-                break
-            # Stop if no improvement limit is passed
-            if self._no_improvement > 1000:
-                break
-            if self._dive and stop:
-                break
-
-    # @profile
     def _find_columns(self):
         "Solves masterproblem and pricing problem."
 
@@ -324,9 +305,9 @@ class VehicleRoutingProblem:
         for vehicle in range(self._vehicle_types):
 
             # Solve pricing problem with randomised greedy algorithm
-            if (self._greedy and not self.time_windows
-                    and not self.distribution_collection
-                    and not self.pickup_delivery):
+            if (self._greedy and not self.time_windows and
+                    not self.distribution_collection and
+                    not self.pickup_delivery):
                 subproblem = self._def_subproblem(duals, vehicle, greedy=True)
                 self.routes, self._more_routes = subproblem.solve(n_runs=20)
 
@@ -342,8 +323,7 @@ class VehicleRoutingProblem:
                         k_shortest_paths,
                     )
                     self.routes, self._more_routes = subproblem.solve(
-                        self._get_time_remaining()
-                    )
+                        self._get_time_remaining())
                     if self._more_routes:
                         break
 
@@ -485,13 +465,9 @@ class VehicleRoutingProblem:
         """
         self._initial_routes = []
         # Run Clarke & Wright if possible
-        if (
-            not self.time_windows
-            and not self.pickup_delivery
-            and not self.distribution_collection
-            and not self.mixed_fleet
-            and not self.periodic
-        ):
+        if (not self.time_windows and not self.pickup_delivery and
+                not self.distribution_collection and not self.mixed_fleet and
+                not self.periodic):
             best_value = 1e10
             best_num_vehicles = 1e10
             for alpha in [x / 10 for x in range(1, 20)]:
@@ -512,17 +488,15 @@ class VehicleRoutingProblem:
                     best_value = alg.best_value
                     best_num_vehicles = len(alg.best_routes)
             logger.info(
-                "Clarke & Wright solution found with value %s and %s vehicles"
-                % (best_value, best_num_vehicles)
-            )
+                "Clarke & Wright solution found with value %s and %s vehicles" %
+                (best_value, best_num_vehicles))
 
             # Run greedy algorithm if possible
-            alg = Greedy(self.G, self.load_capacity, self.num_stops, self.duration)
+            alg = Greedy(self.G, self.load_capacity, self.num_stops,
+                         self.duration)
             alg.run()
-            logger.info(
[…]
