### 2020-06-25 c80507e
Formulate master by columns (#39)
* Initial commit for column-based formulation

- Reorganised main
- Only added set covering constraint by column
- Add method in master to add column

* Remove drop variable from RHS for constraint init

* Finish transfer to column based formulation

* Revert "Finish transfer to column based formulation"

* Finished column formulation

* Remove old column generation method

Co-authored-by: Halvard Storbugt Olsen <HalvardStorbugt.Olsen@sintef.no>
