+++ vrpy/main.py
@@ -81,11 +81,32 @@ class VehicleRoutingProblem:
         self.distribution_collection = distribution_collection
         self.drop_penalty = drop_penalty
         self.fixed_cost = fixed_cost
-        self.num_vehicles = num_vehicles
+        self.num_vehicles = num_vehicles if num_vehicles is not None else []
         self.periodic = periodic
         self.mixed_fleet = mixed_fleet
+        # Parameters for solving
+        self.masterproblem = None
+        self.routes = []
+        self._solver = None
+        self._time_limit = None
+        self._pricing_strategy = None
+        self._exact = None
+        self._cspy = None
+        self._dive = False
+        self._start_time = None
+        self._greedy = None
+        # parameters for column generation stopping criteria
+        self._more_routes = None
+        self._iteration = 0
+        self._no_improvement = 0
+        self._lower_bound = []
+        # Parameters for initial solution
         self._initial_routes = []
         self._preassignments = []
+        # Parameters for final solution
+        self._best_value = None
+        self._best_routes = []
+        self._best_routes_as_graphs = []
 
         # Check if given inputs are consistent
         self._check_vrp()
@@ -149,7 +170,9 @@ class VehicleRoutingProblem:
         Returns:
             float: Optimal solution of MIP based on generated columns
         """
-        # Solving attributes
+
+        # set solving attributes
+        self._more_routes = True
         self._solver = solver
         self._time_limit = time_limit
         self._pricing_strategy = pricing_strategy
@@ -164,14 +187,14 @@ class VehicleRoutingProblem:
             self._initial_routes = initial_routes
 
         # If only one type of vehicle, some formatting is done
-        if self.mixed_fleet == False:
+        if not self.mixed_fleet:
             self._format()
 
         # Pre-processing
         self._pre_solve()
 
         # Initialization
-        self._initialize()
+        self._initialize(solver)
 
         # Column generation
         self._column_generation()
@@ -184,21 +207,8 @@ class VehicleRoutingProblem:
             # Solve as MIP
             # FIXME issue #23
             try:
-                masterproblem_mip = MasterSolvePulp(
-                    self.G,
-                    self._routes_with_node,
-                    self._routes,
-                    self.drop_penalty,
-                    self.num_vehicles,
-                    self.periodic,
-                    self._solver,
-                    self._get_time_remaining(),
-                    relax=False,
-                )
-                (
-                    self._best_value,
-                    self._best_routes_as_graphs,
-                ) = masterproblem_mip.solve()
+                self._best_value, self._best_routes_as_graphs = self.masterproblem.solve(
+                    relax=False)
             except Exception:
                 self._best_value, self._best_routes_as_graphs = (
                     self._lower_bound[-1],
@@ -206,7 +216,7 @@ class VehicleRoutingProblem:
                 )
         # Get dropped nodes
         if self.drop_penalty:
-            self._dropped_nodes = masterproblem_mip.dropped_nodes
+            self._dropped_nodes = self.masterproblem.dropped_nodes
         # Convert best routes into lists of nodes
         self._best_routes_as_node_lists()
         # Schedule routes over time span if Periodic CVRP
@@ -217,7 +227,7 @@ class VehicleRoutingProblem:
                 self.best_routes,
                 self.best_routes_type,
                 self.num_vehicles,
-                self._solver,
+                solver,
             )
             schedule.solve(self._get_time_remaining())
             self._schedule = schedule.routes_per_day
@@ -260,7 +270,7 @@ class VehicleRoutingProblem:
         if self.load_capacity and not self.pickup_delivery:
             self._get_num_stops_upper_bound(self._max_capacity)
 
-    def _initialize(self):
+    def _initialize(self, solver):
[…]
