### 2020-06-29 5c6f193
solve and dive
