+++ tests/unittests/test_toy.py
@@ -57,7 +57,10 @@ class TestsToy:
         """Tests column generation procedure on toy graph
            with stop, capacity and duration constraints
         """
-        prob = VehicleRoutingProblem(self.G, num_stops=3, load_capacity=10, duration=62)
+        prob = VehicleRoutingProblem(self.G,
+                                     num_stops=3,
+                                     load_capacity=10,
+                                     duration=62)
         prob.solve(exact=False)
         assert prob.best_value == 85
         assert set(prob.best_routes_duration.values()) == {41, 62}
@@ -67,7 +70,11 @@ class TestsToy:
         """Tests column generation procedure on toy graph
            with stop, capacity and time_window constraints
         """
-        prob = VehicleRoutingProblem(self.G, num_stops=3, time_windows=True,)
+        prob = VehicleRoutingProblem(
+            self.G,
+            num_stops=3,
+            time_windows=True,
+        )
         prob.solve()
         assert prob.best_value == 80
         assert prob.departure_time[1]["Source"] == 0
@@ -94,21 +101,31 @@ class TestsToy:
     def test_LP_stops_capacity_duration(self):
         """Tests column generation procedure on toy graph"""
         prob = VehicleRoutingProblem(
-            self.G, num_stops=3, load_capacity=10, duration=62,
+            self.G,
+            num_stops=3,
+            load_capacity=10,
+            duration=62,
         )
         prob.solve(cspy=False)
         assert prob.best_value == 85
 
     def test_LP_stops_time_windows(self):
         """Tests column generation procedure on toy graph"""
-        prob = VehicleRoutingProblem(self.G, num_stops=3, time_windows=True,)
+        prob = VehicleRoutingProblem(
+            self.G,
+            num_stops=3,
+            time_windows=True,
+        )
         prob.solve(cspy=False)
         assert prob.best_value == 80
 
     def test_LP_stops_elementarity(self):
         """Tests column generation procedure on toy graph"""
         self.G.add_edge(2, 1, cost=2)
-        prob = VehicleRoutingProblem(self.G, num_stops=3,)
+        prob = VehicleRoutingProblem(
+            self.G,
+            num_stops=3,
+        )
         prob.solve(cspy=False)
         assert prob.best_value == 67
 
@@ -117,9 +134,11 @@ class TestsToy:
     #########
 
     def test_all(self):
-        prob = VehicleRoutingProblem(
-            self.G, num_stops=3, time_windows=True, duration=63, load_capacity=10
-        )
+        prob = VehicleRoutingProblem(self.G,
+                                     num_stops=3,
+                                     time_windows=True,
+                                     duration=63,
+                                     load_capacity=10)
         prob.solve(cspy=False)
         lp_best = prob.best_value
         prob.solve(cspy=True)
@@ -179,7 +198,11 @@ class TestsToy:
         self.G.nodes[5]["demand"] = -10
         self.G.add_edge(2, 5, cost=10)
         self.G.remove_node(1)
-        prob = VehicleRoutingProblem(self.G, load_capacity=15, pickup_delivery=True,)
+        prob = VehicleRoutingProblem(
+            self.G,
+            load_capacity=15,
+            pickup_delivery=True,
+        )
         prob.solve(pricing_strategy="Exact", cspy=False)
         assert prob.best_value == 65
 
@@ -187,7 +210,9 @@ class TestsToy:
         self.G.nodes[1]["collect"] = 12
         self.G.nodes[4]["collect"] = 1
         prob = VehicleRoutingProblem(
-            self.G, load_capacity=15, distribution_collection=True,
+            self.G,
[…]
