+++ tests/benchmarks/test_ortools.py
@@ -45,16 +45,16 @@ class TestsOrTools:
 
     def test_cvrp(self):
         self.prob.load_capacity = 15
-        self.prob.solve(cspy=False, pricing_strategy="PrunePaths")
+        self.prob.solve(cspy=False, pricing_strategy="BestEdges1", dive=True)
         sol_lp = self.prob.best_value
-        self.prob.solve(pricing_strategy="PrunePaths")
+        self.prob.solve(pricing_strategy="BestEdges1")
         sol_cspy = self.prob.best_value
         assert int(sol_lp) == 6208
         assert int(sol_cspy) == 6208
 
     def test_vrptw(self):
         self.prob.time_windows = True
-        self.prob.solve(cspy=False)
+        self.prob.solve(cspy=False, dive=True)
         sol_lp = self.prob.best_value
         self.prob.solve()
         sol_cspy = self.prob.best_value
@@ -64,9 +64,9 @@ class TestsOrTools:
     def test_cvrpsdc(self):
         self.prob.load_capacity = 15
         self.prob.distribution_collection = True
-        self.prob.solve(cspy=False, pricing_strategy="PrunePaths")
+        self.prob.solve(cspy=False, pricing_strategy="BestEdges1", dive=True)
         sol_lp = self.prob.best_value
-        self.prob.solve(pricing_strategy="PrunePaths")
+        self.prob.solve(pricing_strategy="BestEdges1")
         sol_cspy = self.prob.best_value
         assert int(sol_lp) == 6208
         assert int(sol_cspy) == 6208
@@ -80,6 +80,6 @@ class TestsOrTools:
         self.prob.pickup_delivery = True
         self.prob.load_capacity = 10
         self.prob.num_stops = 6
-        self.prob.solve(cspy=False)
+        self.prob.solve(cspy=False, dive=True)
         sol_lp = self.prob.best_value
         assert int(sol_lp) == 5980
