+++ vrpy/newfile.py
@@ -1,0 +1,3 @@
+newfile.py
+
+Python: Show Output
