+++ tests/benchmarks/test_mdvrp_cordeau.py
@@ -10,9 +10,9 @@ class TestsCordeau:
         """
         Cordeau instance p01, 8 vertices only.
         """
-        self.data = DataSet(
-            path="../examples/benchmarks/data/", instance_name="p01", n_vertices=8
-        )
+        self.data = DataSet(path="../examples/benchmarks/data/",
+                            instance_name="p01",
+                            n_vertices=8)
         self.G = self.data.G
 
     def test_setup_instance_name(self):
@@ -28,8 +28,6 @@ class TestsCordeau:
     def test_setup_edges(self):
         assert len(self.G.edges()) == 128
 
-    """
-    Needs some dev
     def test_subproblem_lp_with_initial_routes(self):
         # initial solution
         # ugly, needs more genericity
@@ -37,6 +35,5 @@ class TestsCordeau:
         for v in self.G.nodes():
             if "customer" in self.G.nodes[v]:
                 ini.append(["Source", 51, v, str(51) + "_", "Sink"])
-        self.data.solve(initial_routes=ini, cspy=False)
+        self.data.solve(initial_routes=ini, cspy=False, dive=True)
         assert round(self.data.best_value, 1) == 141.9
-    """

[…]
