### 2020-06-29 7ec58a6
Move time limit param from init to solve method
