### 2020-06-29 #2325 @kakiac
:wave: @jamiehadd, @skadio, @seba-1511, @hausen, @stsievert, @gabrielspadon, @drvinceknight would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
