### 2020-06-29 #2325 @arfon
> Yes, I think we agreed to collaborate, but whedon does not seem to like to have us both, so we have left @bstabler's username for now. Is there a way to add 2 editors?

Got it, thanks. Unfortunately we don't support this.
