### 2020-06-29 #2408 @whedon
[REVIEW]: VRPy: A Python package for solving a range of vehicle routing problems
**Submitting author:** @Kuifje02 (<a href="http://orcid.org/0000-0003-3139-4519">Romain Montagne</a>)
**Repository:** <a href="https://github.com/Kuifje02/vrpy" target ="_blank">https://github.com/Kuifje02/vrpy</a>
**Version:** 0.3.0
