### 2020-06-30 #2408 @bstabler
Hi @Kuifje02, thanks for your submittal.  This is my first review so please be patient as I learn the process.  In addition to the checklist above, I have a few comments at this point:
  - Running the benchmarks requires pandas and matplotlib, but pandas and matplotlib are not listed as dependencies (or at least not on the main page)
  - Running _pytest benchmarks/_ failed many of the tests on my Windows computer?
  - On page 1 one of the paper, change "thus benefits of" to "thus benefits from"
  - I suggest changing the name of the paper to be slightly more specific - "...for solving a range of vehicle routing problems using the column generation approach" - since this seems to be the major contribution?
  - For _Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?_ above, the [API](https://vrpy.readthedocs.io/en/latest/api.html) feels a bit light to me.  Can you add to it?
  - For _Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_, there doesn't seem to be much in the docs related to this.  Can you improve it?
  - For _State of the field: Do the authors describe how this software compares to other commonly-used packages?_, can you add some discussion of other relevant packages?  For example, you mention networkx, cspy, Clp, but not other libraries such as ortools or py-ga-VRPTW.
