### 2020-06-30 Kuifje02/vrpy#43 @torressa
JOSS Review: API
Originally from: https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-651542971

## API
- [x] For _Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?_ above, the [API](https://vrpy.readthedocs.io/en/latest/api.html) feels a bit light to me.  Can you add to it?

- [x] Is vrpy.VehicleRoutingProblem the only "public" interface? I believe so, but if not, please add others to API.html. Btw, maybe add all other classes the prefix "_" to hint the private nature?

Who? All
