### 2020-06-30 Kuifje02/vrpy#44 @torressa
JOSS Review: Paper edits
Originally from: https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-651542971 and https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-655272567

## Paper edits
- [x] On page 1 one of the paper, change "thus benefits of" to "thus benefits from"
- [x] I suggest changing the name of the paper to be slightly more specific - "...for solving a range of vehicle routing problems using the column generation approach" - since this seems to be the major contribution?
- [x] For _Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_, there doesn't seem to be much in the docs related to this.  Can you improve it?
- [x] For _State of the field: Do the authors describe how this software compares to other commonly-used packages?_, can you add some discussion of other relevant packages?  For example, you mention networkx, cspy, Clp, but not other libraries such as ortools or py-ga-VRPTW.

 - [x] Is there any reports/performance you can refer to? If full-blown comparison results are not available, even a quick comment might help the user on what to expect from this library? Is this designed to be competitive in terms of runtime and solution quality? OR, there is no such performance claim but instead, it offers an easy-to-use, unified API for many variants of VRP?

- [x] Please mention in the paper that the library does NOT solve the instances to optimality. See my comment on the performance above.

- [x] You have a nice quick start example in the docs. Maybe add that to the paper?

- [x] "The flexibility and genericity of vrpy is strongly due to ... the relevance of cspy.". I am not sure what you mean here, please consider a rephrase

- [x] The high-level intro to the ColGen approach is nice. For the interested reader who might want to see how different variants of VRP can be formulation in this ColGen framework is there a paper or technical report that you can cite/refer? It is not obvious how all of the variants are modeled, the trick happens in the pricing and a reference of details of each formulation would be nice to have.

- [x] Could you please add a word about other libraries and compare/contrast with this one? Does it solve more variants, why should the reader/user consider this library? For example, you mention examples from the ORTools, a comment on the comparison with ORTools and other libraries would be good for the reader to evaluate.

- [x] Any performance to report and/or refer to? There are many standard VRP benchmarks for this purpose, and I see that you already use Solomon etc.

- [ ] Add some comments on your strategy for testing the library? In particular; testing invalid conditions and testing behavior (optimal value and route)

Who? @torressa @Kuifje02
