### 2020-06-30 Kuifje02/vrpy#42 @torressa
We've already made a start to this to test the diving heuristic (renaming the pricing strategies and adding the lost solve method for Augerat)
We are also thinking of adding a page to the docs with the results.

However, (not urgent!) it'd be good if you could explain how you run the different Augerat instances @Kuifje02:
* Do you run all the sets A, B and P?
* What's the best vrpy set up `greedy=True`, maximum 1 iteration?

That way we can see if diving actually increases the number of optimal solutions found
