### 2020-06-30 5b5e338
solve and dive changes
