+++ tests/benchmarks/test_ortools.py
@@ -43,7 +43,7 @@ class TestsOrTools:
         # Define VRP
         self.prob = VehicleRoutingProblem(self.G)
 
-    def test_cvrp(self):
+    def test_cvrp_dive(self):
         self.prob.load_capacity = 15
         self.prob.solve(cspy=False, pricing_strategy="BestEdges1", dive=True)
         sol_lp = self.prob.best_value
@@ -52,9 +52,50 @@ class TestsOrTools:
         assert int(sol_lp) == 6208
         assert int(sol_cspy) == 6208
 
-    def test_vrptw(self):
+    def test_vrptw_dive(self):
         self.prob.time_windows = True
         self.prob.solve(cspy=False, dive=True)
+        sol_lp = self.prob.best_value
+        self.prob.solve()
+        sol_cspy = self.prob.best_value
+        assert int(sol_lp) == 6528
+        assert int(sol_cspy) == 6528
+
+    def test_cvrpsdc_dive(self):
+        self.prob.load_capacity = 15
+        self.prob.distribution_collection = True
+        self.prob.solve(cspy=False, pricing_strategy="BestEdges1", dive=True)
+        sol_lp = self.prob.best_value
+        self.prob.solve(pricing_strategy="BestEdges1")
+        sol_cspy = self.prob.best_value
+        assert int(sol_lp) == 6208
+        assert int(sol_cspy) == 6208
+
+    def test_pdp_dive(self):
+        # Set demands and requests
+        for (u, v) in PICKUPS_DELIVERIES:
+            self.G.nodes[u]["request"] = v
+            self.G.nodes[u]["demand"] = PICKUPS_DELIVERIES[(u, v)]
+            self.G.nodes[v]["demand"] = -PICKUPS_DELIVERIES[(u, v)]
+        self.prob.pickup_delivery = True
+        self.prob.load_capacity = 10
+        self.prob.num_stops = 6
+        self.prob.solve(cspy=False, dive=True)
+        sol_lp = self.prob.best_value
+        assert int(sol_lp) == 5980
+
+    def test_cvrp(self):
+        self.prob.load_capacity = 15
+        self.prob.solve(cspy=False, pricing_strategy="BestEdges1")
+        sol_lp = self.prob.best_value
+        self.prob.solve(pricing_strategy="BestEdges1")
+        sol_cspy = self.prob.best_value
+        assert int(sol_lp) == 6208
+        assert int(sol_cspy) == 6208
+
+    def test_vrptw(self):
+        self.prob.time_windows = True
+        self.prob.solve(cspy=False)
         sol_lp = self.prob.best_value
         self.prob.solve()
         sol_cspy = self.prob.best_value
@@ -64,7 +105,7 @@ class TestsOrTools:
     def test_cvrpsdc(self):
         self.prob.load_capacity = 15
         self.prob.distribution_collection = True
-        self.prob.solve(cspy=False, pricing_strategy="BestEdges1", dive=True)
+        self.prob.solve(cspy=False, pricing_strategy="BestEdges1")
         sol_lp = self.prob.best_value
         self.prob.solve(pricing_strategy="BestEdges1")
         sol_cspy = self.prob.best_value
@@ -80,6 +121,6 @@ class TestsOrTools:
         self.prob.pickup_delivery = True
         self.prob.load_capacity = 10
         self.prob.num_stops = 6
-        self.prob.solve(cspy=False, dive=True)
+        self.prob.solve(cspy=False)
         sol_lp = self.prob.best_value
         assert int(sol_lp) == 5980
