+++ vrpy/newfile.py
@@ -1,3 +1,0 @@
-newfile.py
-
-Python: Show Output
