### 2020-06-30 #2408 @danielskatz
@bstabler  - I'm getting confused.  Our normal editing workflow is to assign 2 or more reviewers in the pre-review issue, then start the review - this generates the checklists in the review issue.  We should only be assigning reviewers in the review issue in unusual circumstances, and if you do this, you have to manually edit the checklists to make them right.  Look at the first comment in this issue.  It says
> Editor: @kakiac
> Reviewers: @bstabler, @skadio

Is this correct?  If you are a reviewer, why are you doing the editorial work?
