### 2020-06-30 Kuifje02/vrpy#42 @torressa
Fix benchmark tests, still have to rework pandas and matplotlib dependencies.
