### 2020-06-30 760b8e1
Fix Augerat with greedy
