### 2020-06-30 Kuifje02/vrpy#42 @torressa
> Realised I broke the greedy initial solution for the new column based formulation in dev...

Fixed now, I think...
