### 2020-06-30 #2408 @bstabler
Hi @danielskatz, yes that's correct, and thanks for checking in.  @kakiac and I discussed roles with @arfon yesterday and @kakiac will be editor and I'll be a reviewer.  I won't do the editorial work moving forward.  Thanks.
