### 2020-06-30 Kuifje02/vrpy#42 @Kuifje02
> * Running the benchmarks requires pandas and matplotlib, but pandas and matplotlib are not listed as dependencies (or at least not on the main page)

yes for pandas (for reading the data files), but matplotlib can be removed if I am not mistaken. Was there initially to plot the graph. I showed how to do this in the jupyter notebook [here](https://github.com/Kuifje02/vrpy/blob/master/examples/example.ipynb), which will need an update as well.

> * Add benchmarks page to docs with comp results per benchmark

This is definitely a good idea. I was waiting for results to be competitive :D
