### 2020-06-30 #2408 @Kuifje02
> Hi @Kuifje02, thanks for your submittal. This is my first review so please be patient as I learn the process. In addition to the checklist above, I have a few comments at this point:

Hi @bstabler !
Thanks for accepting to review. And thanks for the constructive comments, we'll make the adjustments one by one.
