### 2020-06-30 Kuifje02/vrpy#44 @bstabler
@torressa - note that the community guidelines don't need to go in the paper but could be in a README, a CONTRIBUTING file, or the docs.
