### 2020-06-30 Kuifje02/vrpy#44 @Kuifje02
> * For _State of the field: Do the authors describe how this software compares to other commonly-used packages?_, can you add some discussion of other relevant packages?  For example, you mention networkx, cspy, Clp, but not other libraries such as ortools or py-ga-VRPTW.

Comparisons with ortools are in process. Comparing with py-ga-VRPTW is a great idea! We can run the Solomon instances and compare the three.
