### 2020-07-01 17b8d97
paper edits #44
