### 2020-07-01 2811886
added feasibility check
