+++ vrpy/master_solve_pulp.py
@@ -69,12 +69,15 @@ class MasterSolvePulp(MasterProblemBase):
         constrs = {}
         while depth <= max_depth and len(tabu_list) < max_discrepancy:
             non_integer_vars = list(
-                var for var in relax.variables()
-                if abs(var.varValue - round(var.varValue)) != 0)
+                var
+                for var in relax.variables()
+                if abs(var.varValue - round(var.varValue)) != 0
+            )
             # All non-integer variables not already fixed in this or any
             # iteration of the diving heuristic
             vars_to_fix = [
-                var for var in non_integer_vars
+                var
+                for var in non_integer_vars
                 if var.name not in self._tabu_list and var.name not in tabu_list
             ]
             if vars_to_fix:
@@ -82,21 +85,19 @@ class MasterSolvePulp(MasterProblemBase):
                 # max_discrepancy not violated
 
                 var_to_fix = min(
-                    vars_to_fix,
-                    key=lambda x: abs(x.varValue - round(x.varValue)))
+                    vars_to_fix, key=lambda x: abs(x.varValue - round(x.varValue))
+                )
                 value_to_fix = 1
                 value_previous = var_to_fix.varValue
 
                 name_le = "fix_{}_LE".format(var_to_fix.name)
                 name_ge = "fix_{}_GE".format(var_to_fix.name)
-                constrs[name_le] = pulp.LpConstraint(var_to_fix,
-                                                     pulp.LpConstraintLE,
-                                                     name=name_le,
-                                                     rhs=value_to_fix)
-                constrs[name_ge] = pulp.LpConstraint(var_to_fix,
-                                                     pulp.LpConstraintGE,
-                                                     name=name_ge,
-                                                     rhs=value_to_fix)
+                constrs[name_le] = pulp.LpConstraint(
+                    var_to_fix, pulp.LpConstraintLE, name=name_le, rhs=value_to_fix
+                )
+                constrs[name_ge] = pulp.LpConstraint(
+                    var_to_fix, pulp.LpConstraintGE, name=name_ge, rhs=value_to_fix
+                )
 
                 relax += constrs[name_le]  # add <= constraint
                 relax += constrs[name_ge]  # add >= constraint
@@ -107,8 +108,9 @@ class MasterSolvePulp(MasterProblemBase):
                 # https://github.com/coin-or/pulp/blob/master/pulp/constants.py#L45-L57
                 if not (relax.status != 1):
                     self.prob.extend(constrs)
-                logger.debug("fixed %s with previous value %s", var_to_fix.name,
-                             value_previous)
+                logger.debug(
+                    "fixed %s with previous value %s", var_to_fix.name, value_previous
+                )
             else:
                 break
         self._tabu_list.extend(tabu_list)  # Update global tabu list
@@ -136,9 +138,11 @@ class MasterSolvePulp(MasterProblemBase):
         duals = {}
         # set covering duals
         for node in self.G.nodes():
-            if (node not in ["Source", "Sink"] and
-                    "depot_from" not in self.G.nodes[node] and
-                    "depot_to" not in self.G.nodes[node]):
+            if (
+                node not in ["Source", "Sink"]
+                and "depot_from" not in self.G.nodes[node]
+                and "depot_to" not in self.G.nodes[node]
+            ):
                 constr_name = "visit_node_%s" % node
                 if not relax:
                     duals[node] = self.prob.constraints[constr_name].pi
@@ -150,10 +154,12 @@ class MasterSolvePulp(MasterProblemBase):
             for k in range(len(self.num_vehicles)):
                 if not relax:
[…]
