+++ vrpy/main.py
@@ -204,14 +204,19 @@ class VehicleRoutingProblem:
             self._more_routes = True
             # Initialization
             self._column_generation()
-            self._best_value, self._best_routes_as_graphs = \
-                    self.masterproblem.get_total_cost_and_routes(relax=True)
+            (
+                self._best_value,
+                self._best_routes_as_graphs,
+            ) = self.masterproblem.get_total_cost_and_routes(relax=True)
         else:
             # Solve as MIP
             _, _ = self.masterproblem.solve(
-                relax=False, time_limit=self._get_time_remaining())
-            self._best_value, self._best_routes_as_graphs = \
-                    self.masterproblem.get_total_cost_and_routes(relax=False)
+                relax=False, time_limit=self._get_time_remaining()
+            )
+            (
+                self._best_value,
+                self._best_routes_as_graphs,
+            ) = self.masterproblem.get_total_cost_and_routes(relax=False)
 
         # Get dropped nodes
         if self.drop_penalty:
@@ -258,6 +263,8 @@ class VehicleRoutingProblem:
         self._update_dummy_attributes()
         # Check options consistency
         self._check_consistency()
+        # Check feasibility
+        self._check_feasibility()
         # Lock preassigned routes
         if self._preassignments:
             self._lock()
@@ -278,10 +285,15 @@ class VehicleRoutingProblem:
         # Initial routes are converted to digraphs
         self._convert_initial_routes_to_digraphs()
         # Init master problem
-        self.masterproblem = MasterSolvePulp(self.G, self._routes_with_node,
-                                             self._routes, self.drop_penalty,
-                                             self.num_vehicles, self.periodic,
-                                             solver)
+        self.masterproblem = MasterSolvePulp(
+            self.G,
+            self._routes_with_node,
+            self._routes,
+            self.drop_penalty,
+            self.num_vehicles,
+            self.periodic,
+            solver,
+        )
 
     def _find_columns(self):
         "Solves masterproblem and pricing problem."
@@ -289,26 +301,34 @@ class VehicleRoutingProblem:
         # Solve restricted relaxed master problem
         if self._dive:
             duals, relaxed_cost = self.masterproblem.solve_and_dive(
-                time_limit=self._get_time_remaining())
+                time_limit=self._get_time_remaining()
+            )
         else:
             duals, relaxed_cost = self.masterproblem.solve(
-                relax=True, time_limit=self._get_time_remaining())
+                relax=True, time_limit=self._get_time_remaining()
+            )
         logger.info("iteration %s, %s" % (self._iteration, relaxed_cost))
 
         # One subproblem per vehicle type
         for vehicle in range(self._vehicle_types):
 
             # Solve pricing problem with randomised greedy algorithm
-            if (self._greedy and not self.time_windows and
-                    not self.distribution_collection and
-                    not self.pickup_delivery):
+            if (
+                self._greedy
+                and not self.time_windows
+                and not self.distribution_collection
+                and not self.pickup_delivery
+            ):
                 subproblem = self._def_subproblem(duals, vehicle, greedy=True)
                 self.routes, self._more_routes = subproblem.solve(n_runs=20)
                 # TODO: needs checking as there must be a better way
                 # Update master problem only with new routes
                 if self._more_routes:
-                    for r in (r for r in self.routes
-                              if r.graph["name"] not in self.masterproblem.y):
+                    for r in (
+                        r
+                        for r in self.routes
+                        if r.graph["name"] not in self.masterproblem.y
+                    ):
[…]
