### 2020-07-01 675fd68
skip gencol if all vertices locked closes #50
