### 2020-07-01 b5064ac
unit test for complete lock #50
