### 2020-07-06 c767936
added the csv_table function and modified augeret test and example
