+++ examples/benchmarks/cvrp_augerat.py
@@ -3,10 +3,14 @@ from networkx import relabel_nodes, DiGraph
 import numpy as np
 from pandas import read_csv
 import sys
+import os
+import time
 
 sys.path.append("../../")
-# sys.path.append("../../../cspy")
+sys.path.append("../examples/benchmarks")
+sys.path.append("../../../cspy")
 from vrpy.main import VehicleRoutingProblem
+from examples.benchmarks.report import CsvTableVRP
 
 import logging
 
@@ -15,7 +19,6 @@ logger = logging.getLogger(__name__)
 
 class AugeratNodePosition:
     """Stores coordinates of a node of Augerat's instances (set P)."""
-
     def __init__(self, values):
         # Node ID
         self.name = np.uint32(values[0]).item()
@@ -29,7 +32,6 @@ class AugeratNodePosition:
 
 class AugeratNodeDemand:
     """Stores attributes of a node of Augerat's instances (set P)."""
-
     def __init__(self, values):
         # Node ID
         self.name = np.uint32(values[0]).item()
@@ -46,7 +48,6 @@ class DataSet:
         path (str) : Path to data folder.
         instance_name (str) : Name of instance to read.
     """
-
     def __init__(self, path, instance_name):
 
         # Read vehicle capacity
@@ -112,6 +113,12 @@ class DataSet:
         mapping = dict(zip(before, after))
         self.G = relabel_nodes(self.G, mapping)
 
+        # initialise the table class object
+        self.table = CsvTableVRP(instance_name=instance_name,
+                                 path=path,
+                                 subproblem_type=self.G.graph["name"],
+                                 instance_type="augerat")
+
     def distance(self, u, v):
         """2D Euclidian distance between two nodes.
 
@@ -147,11 +154,17 @@ class DataSet:
             load_capacity=self.max_load,
             num_stops=num_stops,
         )
-        prob.solve(initial_routes=initial_routes,
-                   cspy=cspy,
-                   exact=exact,
-                   time_limit=time_limit,
-                   pricing_strategy=pricing_strategy,
-                   dive=dive,
-                   greedy=greedy)
+        self.table.get_data_from_VRP_instance(
+            prob=prob,
+            dive=dive,
+            subproblem_type=self.G.graph["subproblem"],
+            initial_routes=initial_routes,
+            cspy=cspy,
+            exact=exact,
+            time_limit=time_limit,
+            pricing_strategy=pricing_strategy,
+            greedy=greedy)
+
+        self.table.write_to_file(path="")
+
         self.best_value, self.best_routes = prob.best_value, prob.best_routes
