+++ tests/pytest.ini
@@ -1,4 +1,5 @@
 [pytest]
-#testpaths = tests
+python_files = test_*.py
+testpaths = tests/
 filterwarnings =
     ignore::DeprecationWarning

[…]
