### 2020-07-06 278e5e5
Update resource bounds and custom REFs
