+++ vrpy/subproblem_cspy.py
@@ -1,6 +1,8 @@
-from numpy import array, zeros
 import logging
 import sys
+from math import floor
+
+from numpy import array, zeros
 from networkx import DiGraph, add_path
 
 # sys.path.append("../../cspy")
@@ -39,13 +41,13 @@ class SubProblemCSPY(SubProblemBase):
         # Default lower and upper bounds
         self.min_res = [0] * len(self.resources)
         # Add upper bounds for mono, stops, load and time, and time windows
-        total_demand = sum([self.sub_G.nodes[v]["demand"] for v in self.sub_G.nodes()])
+        total_demand = sum(
+            [self.sub_G.nodes[v]["demand"] for v in self.sub_G.nodes()])
         self.max_res = [
-            len(self.sub_G.nodes()),  # stop/mono
-            total_demand,  # load
-            sum(
-                [self.sub_G.edges[u, v]["time"] for u, v in self.sub_G.edges()]
-            ),  # time
+            floor(len(self.sub_G.nodes()) / 2),  # stop/mono
+            floor(total_demand / 2),  # load
+            sum([self.sub_G.edges[u, v]["time"] for u, v in self.sub_G.edges()
+                ]),  # time
             1,  # time windows
             total_demand,  # pickup
             total_demand,  # deliver
@@ -225,15 +227,31 @@ class SubProblemCSPY(SubProblemBase):
         # load
         new_res[1] += self.sub_G.nodes[j]["demand"]
         # time
-        service_time = self.sub_G.nodes[i]["service_time"]
-        travel_time = self.sub_G.edges[i, j]["time"]
+        # Service times
+        theta_i = self.sub_G.nodes[i]["service_time"]
+        theta_j = self.sub_G.nodes[j]["service_time"]
+        theta_t = self.sub_G.nodes["Sink"]["service_time"]
+        # Travel times
+        travel_time_ij = self.sub_G.edges[i, j]["time"]
+        try:
+            travel_time_jt = self.sub_G.edges[j, "Sink"]["time"]
+        except KeyError:
+            travel_time_jt = 0
+        # Time windows
+        # Lower
         a_j = self.sub_G.nodes[j]["lower"]
+        a_t = self.sub_G.nodes["Sink"]["lower"]
+        # Upper
         b_j = self.sub_G.nodes[j]["upper"]
+        b_t = self.sub_G.nodes["Sink"]["upper"]
 
-        new_res[2] = max(new_res[2] + service_time + travel_time, a_j)
+        new_res[2] = max(new_res[2] + theta_i + travel_time_ij, a_j)
 
         # time-window feasibility resource
-        if not self.time_windows or new_res[2] <= b_j:
+        if not self.time_windows or (
+                new_res[2] <= b_j and \
+                new_res[2] < self.T - a_j - theta_j and \
+                a_t <= new_res[2] + travel_time_jt + theta_t <= b_t):
             new_res[3] = 0
         else:
             new_res[3] = 1
@@ -257,17 +275,32 @@ class SubProblemCSPY(SubProblemBase):
         # load
         new_res[1] += self.sub_G.nodes[i]["demand"]
         # Get relevant service times (thetas) and travel time
+        # Service times
         theta_i = self.sub_G.nodes[i]["service_time"]
         theta_j = self.sub_G.nodes[j]["service_time"]
-        travel_time = self.sub_G.edges[i, j]["time"]
+        theta_s = self.sub_G.nodes["Source"]["service_time"]
+        # Travel times
+        travel_time_ij = self.sub_G.edges[i, j]["time"]
+        try:
+            travel_time_si = self.sub_G.edges["Source", i]["time"]
+        except KeyError:
+            travel_time_si = 0
[…]
