### 2020-07-07 2f9b672
Update CSVTableBase.py
review changes
