### 2020-07-07 a33edaf
fixed comments and created folder for new problem instance data
