### 2020-07-07 5ab953f
Merge PR #51, restructure
