+++ examples/benchmarks/cvrp_augerat.py
@@ -2,23 +2,13 @@ from math import sqrt
 from networkx import relabel_nodes, DiGraph
 import numpy as np
 from pandas import read_csv
-import sys
-import os
-import time
-
-sys.path.append("../../")
-sys.path.append("../examples/benchmarks")
-sys.path.append("../../../cspy")
 from vrpy.main import VehicleRoutingProblem
-from examples.benchmarks.report import CsvTableVRP
-
-import logging
-
-logger = logging.getLogger(__name__)
+from examples.benchmarks.run.csv_table_vrpy import CsvTableVRPy
 
 
 class AugeratNodePosition:
     """Stores coordinates of a node of Augerat's instances (set P)."""
+
     def __init__(self, values):
         # Node ID
         self.name = np.uint32(values[0]).item()
@@ -32,6 +22,7 @@ class AugeratNodePosition:
 
 class AugeratNodeDemand:
     """Stores attributes of a node of Augerat's instances (set P)."""
+
     def __init__(self, values):
         # Node ID
         self.name = np.uint32(values[0]).item()
@@ -48,6 +39,7 @@ class DataSet:
         path (str) : Path to data folder.
         instance_name (str) : Name of instance to read.
     """
+
     def __init__(self, path, instance_name):
 
         # Read vehicle capacity
@@ -114,10 +106,10 @@ class DataSet:
         self.G = relabel_nodes(self.G, mapping)
 
         # initialise the table class object
-        self.table = CsvTableVRP(instance_name=instance_name,
-                                 path=path,
-                                 subproblem_type=self.G.graph["name"],
-                                 instance_type="augerat")
+        self.table = CsvTableVRPy(instance_name=instance_name,
+                                  path=path,
+                                  subproblem_type=self.G.graph["name"],
+                                  instance_type="augerat")
 
     def distance(self, u, v):
         """2D Euclidian distance between two nodes.
@@ -154,7 +146,7 @@ class DataSet:
             load_capacity=self.max_load,
             num_stops=num_stops,
         )
-        self.table.get_data_from_VRP_instance(
+        self.table.get_data_from_VRPy_instance(
             prob=prob,
             dive=dive,
             subproblem_type=self.G.graph["subproblem"],
