### 2020-07-08 #2408 @skadio
Dear authors,

Thank you for introducing an important library. VRP is such a classical optimization problem and having a dedicated package in the Python ecosystem would definitely contribute to further research and experiments.

Overall, the API looks well-designed, the usage example is intuitive, the paper is easy to follow, the documentation and examples are elaborate.

The library passes my initial review, I enjoyed testing and using the software.

My comments below are offered in the spirit of strengthening the library or broadening its appeal to the users.

Best,
Serdar

===

The comments are somewhat in random order. Most of them are minor, and some are "subjective" so feel free to discard.

At a high-level, my main comments would be

*Algorithm* Is there a paper/tech report that shows the formulations of different VRPs using ColGen? The unification approach is very neat and some interested users might want to read more. I say this because VRP, while very applicable, will attract a certain sub-group of OR/Optimization specialists, and then the formulation would be interesting to collect in a document/paper/tech report or individual references for each problem type.

*Performance* Is there any reports/performance you can refer to? If full-blown comparison results are not available, even a quick comment might help the user on what to expect from this library? Is this designed to be competitive in terms of runtime and solution quality? OR, there is no such performance claim but instead it offers an easy-to-use, unified API for many variants of VRP?

*Implementation* There is some hacking with path variables. Some tests are failing because of that. It would be best to switch to relative imports and not alter the system path variable, which might unintended consequences.

## Comments about the manuscript

0. Please mention in the paper that the library does NOT solve the instances to optimality. See my comment on the performance above.

1. You have a nice quick start example in the docs. Maybe add that to the paper?

2. "The flexibility and genericity of vrpy is strongly due to ... the relevance of cspy.". I am not sure what you mean here, please consider a rephrase

3. The high-level intro the ColGen approach is nice. For the interested reader who might want to see how different variants of VRP can be formulation in this ColGen framework is there a paper or technical report that you can cite/refer? It is not obvious how all of the variants is modeled, the trick happens in the pricing and a reference of details of each formulation would be nice to have.

4. Could you please add a word about other libraries and compare/contrast with this one? Does it solve more variants, why should the reader/user consider this library? For example, you mention examples from the ORTools, a comment on the comparison with ORTools and other libraries would be good for the reader to evaluate.

5. Any performance to report and/or refer to? There are many standard VRP benchmarks for this purpose, and I see that you already use Solomon etc.

6. Add some comments on your strategy for testing the library? In particular; testing invalid conditions and testing behavior (optimal value and route)

## Comments about the Library

1. Git clone failed for me. It might as well be a problem with my network/firewall setup. @bstabler can you please check that. I downloaded the zip instead.

2. Out of the box, running tests in unittests/ folder from PyCharm (on Windows) worked for me. Test_toy took some time to run but maybe that's expected

3. Out of the box, all tests in benchmarks/ fail. All errors, except the one in test_ortools, are the same:

>       with open(path + instance_name) as fp:
E       FileNotFoundError: [Errno 2] No such file or directory: '../examples/benchmarks/data/P-n16-k8.vrp'

..\..\examples\benchmarks\cvrp_augerat.py:53: FileNotFoundError
****

You seem to have an issue with setting path variables. Please check.

4. test_ortools error is

    from ortools.data import (
E   ModuleNotFoundError: No module named 'ortools'

Again, a similar issue. It can't find the ortools folder which is under examples/

5. Related to this. I suggest renaming "ortools" folder. The current import statements read as if you have a dependency on the ortools library. It's not obvious that it's just a folder in the upper level.

6. Some examples are using numpy.matrix(). It is not recommended to use. Please consider regular numpy ndarrays.

7. In test_toy.py remove from pytest import raises --not used. same for test_greedy

8. cvrp_augeratpy remove from vrpy.main import VehicleRoutingProblem --not used

9. In benchmarks folder, some files have a main(), some files dont. Why?

10. Running mdvrp_cordeau.py fails with
raise KeyError("Node %s missing from initial solution." % v)
KeyError: 'Node 52 missing from initial solution.'

11. In examples/ folder, you typically print best value and best routes. Please also ASSERT them. I can run them, I don't know if the solution is correct or not.

BTW, why is the separate between test and examples. Why not add these examples to test and
12. Btw, is there a way to turn off logging from underlying solvers like cbc etc. There is a lot printouts that is not directly related to this library.

I am not convinced by the structure of the "examples" folder. My expectation as a user would be to see examples of how to solve different variants of VRP problems. The "ortools" folder does not serve that purpose. The user needs to know what ortools is and that what is meant by that folder is examples from ortools.

In my mind; examples folder looks like

examples/
|_ vrp
|_ vrp_tw
|_ capacited_vrp
|_ capacited_vrp_xxx (xxx is some benchmark name)
|_ multi_depo_cvrp
|_ pickup_delivery

examples.ipync is nice.. but please add ASSERTs along with prints()

13. You might want to add LICENSE information as a header on top of every source code file. Maybe also add License section to README?

14. Thinking out loud comment: if you rename main.py to vrp or vehicle_routing, the imports would read:

from vrpy.vrp import VehicleRoutingProblem
from vrpy.vehicle_routing import VehicleRoutingProblem

which might read a bit more relevant than "main"

15. There is a lot of mingling with path variable to append sys.path for folders. I don't think this is necessary and complicates the code/usage --since you need to track what's in the path. I don't see anything special in your library structure that would require altering with path variables. Cannot we resolve these import issues without hacking into the path?

## Comments about API

1. Is vrpy.VehicleRoutingProblem the only "public" interface? I believe so, but if not, please add others to API.html. Btw, maybe refactor all other classes with prefix "_" to hint the private nature?

## Comments about Docs

1. Math Background sections have a lot of relevant references, but most are not cited in the text.

2. The section "Why use the cspy library?" is somewhat misleading as part of the "VRPy" library. We are not directly using the cspy here. Maybe call if "Solving the Pricing using cspy" or sth similar?

3. In Index.html it reads "solving Vehicle Routing Problems (VRP)". What you probably mean is solving "instances" of different types of Vehicle Routing Problems. Here you should mention that the library is NOT guaranteed to find optimal solutions.

## Comments about Installation

1. pip install worked as expected, after installing requirements.
2. I was able to import and run the quick start example with no issues.
