### 2020-07-08 #2408 @skadio
@Kuifje02 @bstabler @kakiac

I am done with my 1st pass of the review. Overall, I am pleased with the library, detailed comments above.
