### 2020-07-08 Kuifje02/vrpy#42 @torressa
Most of the import errors have been sorted in the `Halvaros/dev` branch, will check when merged into dev.
And I think the Cordeau has also been removed in the `joss` branch.
