### 2020-07-08 Kuifje02/vrpy#52 @torressa
JOSS Review: Examples
Originally from: https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-655272567

# Structure of `examples/`

## Folder structure

btw, I am not convinced by the structure of the "examples" folder. My expectation as a user would be to see examples of how to solve different variants of VRP problems. The benchmark folder is somewhat okay but the "ortools" folder does not serve that purpose. The user needs to know what ortools is in the first place, and that what is meant by that folder is not really the library but the examples from ortools.

In my mind; examples folder would look like sth like this:

examples/
|_ vrp
|_ vrp_tw
|_ capacited_vrp
|_ capacited_vrp_xxx (xxx is some benchmark name)
|_ multi_depo_cvrp
|_ pickup_delivery

examples.ipync is nice.. but please add ASSERTs along with prints() so that we can verify correctness.

## Separation between tests and examples

In examples/ folder, you typically print the best value and best route. Please also ASSERT them. I can run them, but I don't know if the solution is correct or not.

BTW, why is the separation between tests and examples. Why not add these examples to test and assert them.
