### 2020-07-08 Kuifje02/vrpy#42 @torressa
We need to have a discussion about the structure point. Whether moving the tests in tests/benchmarks into the examples folders and the other stuff in #52
