### 2020-07-08 Kuifje02/vrpy#42 @Kuifje02
I am thinking maybe it would make sense to have everything related to benchmarks in a one independent folder, and only usage examples in the examples folder. So as suggested in #52, maybe :

examples/
|_ vrp
|_ vrp_tw
|_ capacited_vrp
|_ multi_depo_cvrp
|_ pickup_delivery
|_ etc

benchmarks/
|_data/
|_cvrp_augerat
|_cvrptw_solomon
 ++ @Halvaros' new scripts and files
++ maybe the tests that are in the tests/benchmarks right now ? (not sure)
