### 2020-07-08 Kuifje02/vrpy#54 @Kuifje02
JOSS Review: Docs
# Comments about Docs

- [x] Math Background sections have a lot of relevant references, but most are not cited in the text.

- [x] The section "Why use the cspy library?" is somewhat misleading in the doc section of the "VRPy" library. We are not directly using the cspy here. Maybe call if "Solving the Pricing using cspy" or sth similar?

- [x] In Index.html it reads "solving Vehicle Routing Problems (VRP)". What you probably mean is solving "instances" of different types of Vehicle Routing Problems. Here you should definitely mention that the library is NOT guaranteed to find optimal solutions, and maybe a word on the solution quality of what it typically achieves.

- [x] Add performance profiles to docs.
