### 2020-07-08 Kuifje02/vrpy#42 @torressa
That split makes more sense! Moving the test files there also makes sense.

benchmarks/
|_data/
|_run/ <- scripts for reproducing results (writing table and running different instances)
|_tests/ <- Put `tests/benchmarks/` here
|_cvrp_augerat.py
|_cvrptw_solomon.py
