### 2020-07-08 276b059
Included the run method and made minor changes in main
