+++ vrpy/subproblem_lp.py
@@ -1,7 +1,7 @@
 from networkx import DiGraph, negative_edge_cycle, shortest_path
 import pulp
 import logging
-from vrpy.subproblem import SubProblemBase
+from .subproblem import SubProblemBase
 
 logger = logging.getLogger(__name__)
 
@@ -13,13 +13,14 @@ class SubProblemLP(SubProblemBase):
 
     Inherits problem parameters from `SubproblemBase`
     """
-
     def __init__(self, *args, solver):
         super(SubProblemLP, self).__init__(*args)
         # create problem
         self.prob = pulp.LpProblem("SubProblem", pulp.LpMinimize)
         # flow variables
-        self.x = pulp.LpVariable.dicts("x", self.sub_G.edges(), cat=pulp.LpBinary)
+        self.x = pulp.LpVariable.dicts("x",
+                                       self.sub_G.edges(),
+                                       cat=pulp.LpBinary)
         self.solver = solver
 
     # @profile
@@ -38,10 +39,10 @@ class SubProblemLP(SubProblemBase):
         logger.debug("Solving subproblem using LP")
         logger.debug("Status: %s" % pulp.LpStatus[self.prob.status])
         logger.debug("Objective %s" % pulp.value(self.prob.objective))
-        if (
-            pulp.value(self.prob.objective) is not None
-            and pulp.value(self.prob.objective) < -(10 ** -3)
-        ) or (exact == False and pulp.LpStatus[self.prob.status] in ["Optimal", ""]):
+        if (pulp.value(self.prob.objective) is not None
+                and pulp.value(self.prob.objective) < -(10**-3)) or (
+                    exact == False
+                    and pulp.LpStatus[self.prob.status] in ["Optimal", ""]):
             more_routes = True
             self._add_new_route()
             return self.routes, more_routes
@@ -64,10 +65,10 @@ class SubProblemLP(SubProblemBase):
         new_route.graph["cost"] = self.total_cost
         new_route.graph["vehicle_type"] = self.vehicle_type
         self.routes.append(new_route)
-        logger.debug(
-            "new route %s %s" % (route_id, shortest_path(new_route, "Source", "Sink"))
-        )
-        logger.debug("new route reduced cost %s" % pulp.value(self.prob.objective))
+        logger.debug("new route %s %s" %
+                     (route_id, shortest_path(new_route, "Source", "Sink")))
+        logger.debug("new route reduced cost %s" %
+                     pulp.value(self.prob.objective))
         logger.debug("new route cost = %s" % self.total_cost)
         # for (i, j) in new_route.edges():
         #    print(i, j, self.sub_G.edges[i, j])
@@ -85,32 +86,30 @@ class SubProblemLP(SubProblemBase):
 
     def _formulate(self):
         # minimize reduced cost
-        self.prob += pulp.lpSum(
-            [
-                self.sub_G.edges[i, j]["weight"] * self.x[(i, j)]
-                for (i, j) in self.sub_G.edges()
-            ]
-        )
+        self.prob += pulp.lpSum([
+            self.sub_G.edges[i, j]["weight"] * self.x[(i, j)]
+            for (i, j) in self.sub_G.edges()
+        ])
         # flow balance
         for v in self.sub_G.nodes():
             if v not in ["Source", "Sink"]:
                 in_flow = pulp.lpSum(
-                    [self.x[(i, v)] for i in self.sub_G.predecessors(v)]
-                )
+                    [self.x[(i, v)] for i in self.sub_G.predecessors(v)])
                 out_flow = pulp.lpSum(
-                    [self.x[(v, j)] for j in self.sub_G.successors(v)]
-                )
+                    [self.x[(v, j)] for j in self.sub_G.successors(v)])
                 self.prob += in_flow == out_flow, "flow_balance_%s" % v
[…]
