### 2020-07-09 #2408 @kakiac
:wave @bstabler @skadio Thanks all for your hard work and your extensive feedback! 🥇

@Kuifje02 - would you please address their comments, and upload a revised version on your submission in https://github.com/Kuifje02/vrpy/tree/master/paper?
