### 2020-07-09 #2408 @kakiac
LIst of JOSS review related issues:

- https://github.com/Kuifje02/vrpy/issues/54
- https://github.com/Kuifje02/vrpy/issues/53
- https://github.com/Kuifje02/vrpy/issues/52
- https://github.com/Kuifje02/vrpy/issues/44
- https://github.com/Kuifje02/vrpy/issues/43
- https://github.com/Kuifje02/vrpy/issues/42
