### 2020-07-09 b26a29b
Changes to the run function
