+++ examples/benchmarks/run/csv_table_vrpy.py
@@ -7,8 +7,8 @@ class CsvTableVRPy(CsvTableBase):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-    def write_to_file(self, path='report'):
-        super().write_to_file(path)
+    def write_to_file(self, path_to):
+        super().write_to_file(path_to=path_to)
 
     def _find_optimal(self):
         if self.instance_type == "augerat":
@@ -24,6 +24,37 @@ class CsvTableVRPy(CsvTableBase):
             self.best_known_solution = None
 
     def get_data_from_VRPy_instance(self,
+                                    best_known_solution=None,
+                                    dive=None,
+                                    pricing_strategy=None,
+                                    best_value=None,
+                                    lower_bound=None,
+                                    comp_time=None,
+                                    cspy=None,
+                                    greedy=None):
+
+        if best_known_solution is None:
+            self._find_optimal()
+        else:
+            self.best_known_solution = best_known_solution
+
+        self.dive = dive if not dive is None else False
+        self.pricing_strategy = pricing_strategy
+        self.subproblem_type = "cspy" if cspy else "lp"
+
+        self.comp_time = comp_time
+        self.integrality_gap = (self.upper_bound -
+                                self.lower_bound) / self.lower_bound * 100
+
+        if not self.best_known_solution is None:
+            self.optimality_gap = (self.upper_bound - self.best_known_solution
+                                   ) / self.best_known_solution * 100
+            self.optimal = (self.optimality_gap == 0)
+        else:
+            self.optimality_gap = "Unknown"
+            self.optimal = "Unknown"
+
+    """ def get_data_from_VRPy_instance(self,
                                     prob=None,
                                     best_known_solution=None,
                                     subproblem_type=None,
@@ -57,4 +88,4 @@ class CsvTableVRPy(CsvTableBase):
             self.optimal = (self.optimality_gap == 0)
         else:
             self.optimality_gap = "Unknown"
-            self.optimal = "Unknown"
+            self.optimal = "Unknown" """
