### 2020-07-09 ab60e4d
paper edits #44
