### 2020-07-09 a5fcbb0
paper edits, small typos #44
