### 2020-07-09 Kuifje02/vrpy#44 @Kuifje02
> * Add some comments on your strategy for testing the library? In particular; testing invalid conditions and testing behavior (optimal value and route)

Not sure I understand this one
