### 2020-07-09 Kuifje02/vrpy#44 @Kuifje02
Most suggestions from https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-655272567 have been taken into account, just need @torressa to validate modifications.

Also, I hope the paper is not too long now.
