### 2020-07-09 7c96975
paper edits, small typos #44
