### 2020-07-09 Kuifje02/vrpy#54 @Kuifje02
> * Math Background sections have a lot of relevant references, but most are not cited in the text.

If I am not mistaken most (all?) references are cited, but not all of them are in the _Mathematical background_ section. Would it make more sense to add a _Bibliography_ section (as in the [NetworkX](https://networkx.github.io/documentation/stable/) docs) ? Why not
