+++ docs/benchmarks.rst
@@ -1,0 +1,4 @@
+.. _benchmarks:
+
+Performance profiles
+====================
