+++ docs/bibliography.rst
@@ -2,3 +2,4 @@ Bibliography
 ------------
 
 .. bibliography:: refs.bib
+   :cited:
