+++ docs/refs.bib
@@ -1,33 +1,25 @@
-@article{costa2019exact,
-  title={Exact branch-price-and-cut algorithms for vehicle routing},
-  author={Costa, Luciano and Contardo, Claudio and Desaulniers, Guy},
-  journal={Transportation Science},
-  volume={53},
-  number={4},
-  pages={946--985},
-  year={2019},
-  publisher={INFORMS}
+@Misc{johnjforrest_2020,
+  author       = {johnjforrest and
+                  Stefan Vigerske and
+                  Ted Ralphs and
+                  Lou Hafer and
+                  jpfasano and
+                  Haroldo Gambini Santos and
+                  Matthew Saltzman and
+                  h-i-gassmann and
+                  Bjarni Kristjansson and
+                  Alan King},
+  title        = {coin-or/Clp: Version 1.17.6},
+  month        = apr,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {releases/1.17.6},
+  doi          = {10.5281/zenodo.3748677},
+  url          = {https://doi.org/10.5281/zenodo.3748677}
 }
 
-@incollection{irnich2005shortest,
-  title={Shortest path problems with resource constraints},
-  author={Irnich, Stefan and Desaulniers, Guy},
-  booktitle={Column generation},
-  pages={33--65},
-  year={2005},
-  publisher={Springer}
-}
 
-@article{righini2006symmetry,
-  title={Symmetry helps: Bounded bi-directional dynamic programming for the elementary shortest path problem with resource constraints},
-  author={Righini, Giovanni and Salani, Matteo},
-  journal={Discrete Optimization},
-  volume={3},
-  number={3},
-  pages={255--273},
-  year={2006},
-  publisher={Elsevier}
-}
+
 
 @Misc{cspy,
   author = {Torres Sanchez, David},
@@ -122,32 +114,6 @@
   publisher={Elsevier}
 }
 
-@Misc{johnjforrest_2020,
-  author       = {johnjforrest and
-                  Stefan Vigerske and
-                  Ted Ralphs and
-                  Lou Hafer and
-                  jpfasano and
-                  Haroldo Gambini Santos and
-                  Matthew Saltzman and
-                  h-i-gassmann and
-                  Bjarni Kristjansson and
-                  Alan King},
-  title        = {coin-or/Clp: Version 1.17.6},
-  month        = apr,
-  year         = 2020,
-  publisher    = {Zenodo},
-  version      = {releases/1.17.6},
-  doi          = {10.5281/zenodo.3748677},
-  url          = {https://doi.org/10.5281/zenodo.3748677}
-}
 
-@incollection{desrosiers1988shortest,
-  title={The shortest path problem for the construction of vehicle routes with pick-up, delivery and time constraints},
-  author={Desrosiers, Jacques and Dumas, Yvan},
-  booktitle={Advances in Optimization and Control},
-  pages={144--157},
-  year={1988},
-  publisher={Springer}
-}
+
 
