### 2020-07-09 Kuifje02/vrpy#42 @torressa
Working on the restructuring and path fixing thing... Big commit dropping soon
