### 2020-07-09 af6cd5f
benchmarks in docs #54 #49
