### 2020-07-09 Kuifje02/vrpy#54 @Kuifje02
JOSS follow up https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-656075202
