### 2020-07-09 33fa346
Restructure and rework of benchmarks (#42)
- csv table writer
- benchmark runner
- deleted .solve methods from benchmarks, too much hassle for
consistency
- moved examples/benchmarks/ -> benchmarks/
- moved tests/benchmarks/ -> benchmarks/tests/
