+++ vrpy/subproblem_lp.py
@@ -13,6 +13,7 @@ class SubProblemLP(SubProblemBase):
 
     Inherits problem parameters from `SubproblemBase`
     """
+
     def __init__(self, *args, solver):
         super(SubProblemLP, self).__init__(*args)
         # create problem
@@ -39,10 +40,10 @@ class SubProblemLP(SubProblemBase):
         logger.debug("Solving subproblem using LP")
         logger.debug("Status: %s" % pulp.LpStatus[self.prob.status])
         logger.debug("Objective %s" % pulp.value(self.prob.objective))
-        if (pulp.value(self.prob.objective) is not None
-                and pulp.value(self.prob.objective) < -(10**-3)) or (
-                    exact == False
-                    and pulp.LpStatus[self.prob.status] in ["Optimal", ""]):
+        if (pulp.value(self.prob.objective) is not None and
+                pulp.value(self.prob.objective) < -(10**-3)) or (
+                    exact == False and
+                    pulp.LpStatus[self.prob.status] in ["Optimal", ""]):
             more_routes = True
             self._add_new_route()
             return self.routes, more_routes
@@ -81,8 +82,14 @@ class SubProblemLP(SubProblemBase):
         elif self.solver == "cplex":
             self.prob.solve(pulp.CPLEX_CMD(msg=0, timelimit=time_limit))
         elif self.solver == "gurobi":
-            gurobi_options = [("TimeLimit", time_limit)]
-            self.prob.solve(pulp.GUROBI_CMD(options=gurobi_options))
+            gurobi_options = []
+            # Only specify time limit if given (o.w. errors)
+            if time_limit is not None:
+                gurobi_options.append((
+                    "TimeLimit",
+                    time_limit,
+                ))
+            self.prob.solve(pulp.GUROBI(msg=0, options=gurobi_options))
 
     def _formulate(self):
         # minimize reduced cost
