### 2020-07-09 1f2edfb
Joss (#55)
* paper edits #44

* paper edits #44

* paper edits, small typos #44

* paper edits, small typos #44

* paper edits, small typos #44

Co-authored-by: Romain MONTAGNÉ <romain.montagne@fr.eurodecision.com>
