+++ paper/paper.bib
@@ -203,3 +203,43 @@
   year={2007},
   publisher={Wiley Online Library}
 }
+
+@article{sadykov2019primal,
+  title={Primal heuristics for branch and price: The assets of diving methods},
+  author={Sadykov, Ruslan and Vanderbeck, Fran{\c{c}}ois and Pessoa, Artur and Tahiri, Issam and Uchoa, Eduardo},
+  journal={INFORMS Journal on Computing},
+  volume={31},
+  number={2},
+  pages={251--267},
+  year={2019},
+  publisher={INFORMS}
+}
+
+@article{solomon1987algorithms,
+  title={Algorithms for the vehicle routing and scheduling problems with time window constraints},
+  author={Solomon, Marius M},
+  journal={Operations research},
+  volume={35},
+  number={2},
+  pages={254--265},
+  year={1987},
+  publisher={Informs}
+}
+
+@article{pessoa2018automation,
+  title={Automation and combination of linear-programming based stabilization techniques in column generation},
+  author={Pessoa, Artur and Sadykov, Ruslan and Uchoa, Eduardo and Vanderbeck, Fran{\c{c}}ois},
+  journal={INFORMS Journal on Computing},
+  volume={30},
+  number={2},
+  pages={339--360},
+  year={2018},
+  publisher={INFORMS}
+
+@phdthesis{augerat1995approche,
+  title={Approche poly{\`e}drale du probl{\`e}me de tourn{\'e}es de v{\'e}hicules},
+  author={Augerat, Philippe},
+  year={1995},
+  school={Institut National Polytechnique de Grenoble-INPG}
+}
+
