### 2020-07-09 1a09985
Revert "Joss (#55)"
This reverts commit 1f2edfb62001a9d980824e8f15d2199ba8dc64ec.
