+++ paper/paper.md
@@ -33,7 +33,7 @@ Mathematicians have started tackling VRPs since 1959 [@dantzig1959truck]. Ever s
 
 # Features
 
-``vrpy`` is a Python package that offers an easy-to-use, unified API for many variants of VRP including:
+``vrpy`` is a Python package that allows one to solve variants of the VRP including:
 
 -   the Capacitated VRP (CVRP) [@laporte2007you;@baldacci2010exact],
 -   the CVRP with resource constraints [@laporte1985optimal],
@@ -44,78 +44,25 @@ Mathematicians have started tackling VRPs since 1959 [@dantzig1959truck]. Ever s
 
 For each of these variants, it is possible to i/ set initial routes for the search (if one already has a solution at hand and wishes to improve it) ii/ lock routes (if part of the solution is imposed and must not be optimized) iii/ drop nodes (ignore a customer at the cost of a penalty).
 
-``vrpy`` is built upon the well known *NetworkX* library [@hagberg2008exploring] and thus benefits from a user friendly API, as shown in the following quick start example:
-
-```python
-from networkx import DiGraph
-from vrpy import VehicleRoutingProblem
-
-# Define the network
-G = DiGraph()
-G.add_edge("Source",1,cost=1,time=2)
-G.add_edge("Source",2,cost=2,time=1)
-G.add_edge(1,"Sink",cost=0,time=2)
-G.add_edge(2,"Sink",cost=2,time=3)
-G.add_edge(1,2,cost=1,time=1)
-G.add_edge(2,1,cost=1,time=1)
-
-# Define the customers demands
-G.nodes[1]["demand"] = 5
-G.nodes[2]["demand"] = 4
-
-# Define the Vehicle Routing Problem
-prob = VehicleRoutingProblem(G, load_capacity=10, duration=5)
-
-# Solve and display solution value
-prob.solve()
-print(prob.best_value)
-3
-print(prob.best_routes)
-{1: ["Source",2,1,"Sink"]}
-```
-
-# State of the field
-
-Although the VRP is a classical optimization problem, to our knowledge there is only one dedicated package in the Python ecosystem that is able to solve such a range of VRP variants: the excellent ``OR-Tools`` (Google) routing library [@ortools], released for the first time in 2014. To be precise, the core algorithms are implemented in C++, but the library provides a wrapper in Python. Popular and efficient, it is a reference for ``vrpy``, both in terms of features and performance. The current version of ``vrpy`` is able to handle the same variants as OR-Tools (mentioned in the previous section).
-
-Performance-wise, ``vrpy`` ambitions to be competitive with ``OR-Tools`` eventually, at least in terms of solution quality. For the moment, benchmarks (available in the repository) for the CVRP on the set of Augerat instances ([@augerat1995approche]) show promising results: in the performance profile in Figure 1 below, one can see that nearly the same number of instances are solved within 10 seconds with the same relative error with respect to the best known solution (42\% for ``vrpy``, 44\% for ``OR-Tools``).
-
-| ![Performance profile](cvrp_performance_profile.png) |
-| :--------------------------------------------------: |
-|         *Figure 1: CVRP Performance profile*         |
-
-We do not claim to outperform ``OR-Tools``, but aim to have results of the same order of magnitude as the package evolves, as there is still much room for improvement (see Section *Future Work* below). On the other hand, we are confident that the user friendly and intuitive API will help students, researchers and more generally the operational research community solve instances of vehicle routing problems of small to medium size, perhaps more easily than with the existing software.
-
[…]
