### 2020-07-09 82f7968
Increase time limit test total for CI
