### 2020-07-10 #2408 @bstabler
@skadio - git clone works for me on Windows.
