### 2020-07-10 516ae97
Update running script
