+++ benchmarks/run.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python
-
 import os
 from time import time
 import argparse
 from itertools import product
+from logging import getLogger
 from multiprocessing import Pool, cpu_count
 
 from networkx import relabel_nodes, DiGraph
@@ -14,83 +14,155 @@ from benchmarks.utils.csv_table import CsvTable
 
 from vrpy import VehicleRoutingProblem
 
+logger = getLogger("run")
+
 parser = argparse.ArgumentParser(description='Run benchmarks.')
+
+parser.add_argument('--input_folder',
+                    '-dir',
+                    type=str,
+                    default="benchmarks/data/",
+                    dest="INPUT_FOLDER",
+                    help='Top folder where the instances are located' +
+                    ' Default: benchmarks/data/')
 parser.add_argument('--instance_types',
+                    '-i',
                     nargs='+',
                     dest="INSTANCE_TYPES",
-                    help='Type of instance to run: crvp, cvrptw')
+                    default=["cvrp"],
+                    help='Type of instance to run: crvp, cvrptw.' +
+                    ' Note same name as folder in data. Default: cvrp')
 parser.add_argument('--time_limit',
                     type=int,
-                    default=100,
+                    default=10,
                     dest="TIME_LIMIT",
-                    help='Time limit for each instance in seconds')
+                    help='Time limit for each instance in seconds.' +
+                    ' Default: 10')
+parser.add_argument('--series',
+                    '-s',
+                    action='store_true',
+                    dest="SERIES",
+                    help='To run the benchmarks in series or in parallel.' +
+                    ' Default: parallel')
+parser.add_argument('--cpu-count',
+                    '-cpu',
+                    type=int,
+                    default=0,
+                    dest="CPU_COUNT",
+                    help='Number of cpus to use. Default: all avaiable.')
+parser.add_argument('--exploration',
+                    '-e',
+                    action='store_false',
+                    dest="PERFORMANCE",
+                    help='To run the benchmarks in performance mode (default)' +
+                    ' or in exploration mode. Exploration mode runs' +
+                    ' different solver parameters')
+
 args = parser.parse_args()
-# Set global vars
+
+# Set vars from arguments
+INPUT_FOLDER = args.INPUT_FOLDER
 INSTANCE_TYPES = args.INSTANCE_TYPES
 TIME_LIMIT = args.TIME_LIMIT
+SERIES = args.SERIES
+CPU_COUNT = cpu_count() if not args.CPU_COUNT else args.CPU_COUNT
+PERFORMANCE = args.PERFORMANCE
+# Perfomance set up
+PERFORMANCE_SOLVER_PARAMS = {
+    'cvrp': {
+        'dive': False,
+        'greedy': True,
+        'cspy': False,
+        'pricing_strategy': 'BestEdges1'
+    },
+    'cvrptw': {
+        'dive': False,
+        'greedy': True,
+        'cspy': True,
+        'pricing_strategy': 'BestEdges1'
+    }
+}
 
 
-def run_series(input_folder=None, time_limit=100):
+def run_series():
     """
     Iterates through all problem instances and creates csv table
     in a new folder `benchmarks/results/` in series
     """
-    if input_folder is None:
-        input_folder = "benchmarks/data/"
 
     for instance_type in INSTANCE_TYPES:
-        path_to_instance_type = input_folder + instance_type + "/"
-        for file in os.listdir(path_to_instance_type):
-            filename = os.fsdecode(file)
-            if filename.endswith(".vrp"):
-                for dive in [True, False]:
-                    for cspy in [True, False]:
-                        for pricing_strategy in [
-                                "BestPaths", "BestEdges1", "BestEdges2", "Exact"
-                        ]:
-                            for greedy in [True, False]:
[…]
