+++ benchmarks/benchmarks.rst
@@ -1,16 +1,0 @@
-
-Benchmarks
-==========
-
-
-Results
-~~~~~~~
-
-
-Reproducing results
-~~~~~~~~~~~~~~~~~~~
-
-To run all the Augerat instances do
-
-.. code:: bash
-    python3 -m benchmarks.run --instance_types cvrp

[…]
