### 2020-07-10 Kuifje02/vrpy#42 @torressa
Last one:

>  Running pytest benchmarks/ failed many of the tests on my Windows computer? The examples worked and the tests passed.

Anyone on a \ Windows\ machine pls
