### 2020-07-10 Kuifje02/vrpy#42 @torressa
Still have to maybe move the ortools related tests to the examples folder maybe
