### 2020-07-10 b2556ef
Move from os -> pathlib (OS safer paths)
