+++ benchmarks/run.py
@@ -1,10 +1,11 @@
 #!/usr/bin/env python
-import os
 from time import time
 import argparse
 from itertools import product
 from logging import getLogger
 from multiprocessing import Pool, cpu_count
+from pathlib import Path
+from typing import List, Dict, Union
 
 from networkx import relabel_nodes, DiGraph
 
@@ -61,14 +62,14 @@ parser.add_argument('--exploration',
 args = parser.parse_args()
 
 # Set vars from arguments
-INPUT_FOLDER = args.INPUT_FOLDER
-INSTANCE_TYPES = args.INSTANCE_TYPES
-TIME_LIMIT = args.TIME_LIMIT
-SERIES = args.SERIES
-CPU_COUNT = cpu_count() if not args.CPU_COUNT else args.CPU_COUNT
-PERFORMANCE = args.PERFORMANCE
+INPUT_FOLDER: Path = Path(args.INPUT_FOLDER)
+INSTANCE_TYPES: List[str] = args.INSTANCE_TYPES
+TIME_LIMIT: int = args.TIME_LIMIT
+SERIES: bool = args.SERIES
+CPU_COUNT: int = cpu_count() if not args.CPU_COUNT else args.CPU_COUNT
+PERFORMANCE: bool = args.PERFORMANCE
 # Perfomance set up
-PERFORMANCE_SOLVER_PARAMS = {
+PERFORMANCE_SOLVER_PARAMS: Dict[str, Dict[str, Union[bool, str]]] = {
     'cvrp': {
         'dive': False,
         'greedy': True,
@@ -85,39 +86,33 @@ PERFORMANCE_SOLVER_PARAMS = {
 
 
 def run_series():
-    """
-    Iterates through all problem instances and creates csv table
+    """Iterates through all problem instances and creates csv table
     in a new folder `benchmarks/results/` in series
     """
-
     for instance_type in INSTANCE_TYPES:
-        path_to_instance_type = os.path.join(INPUT_FOLDER, instance_type)
-        for root, _, files in os.walk(path_to_instance_type):
-            for f in files:
-                path_to_instance = os.path.join(root, f)
-                if PERFORMANCE:
-                    _run_single_problem(
-                        path_to_instance,
-                        **PERFORMANCE_SOLVER_PARAMS[instance_type])
-                else:
-                    for dive in [True, False]:
-                        for cspy in [True, False]:
-                            for pricing_strategy in [
-                                    "BestPaths", "BestEdges1", "BestEdges2",
-                                    "Exact"
-                            ]:
-                                for greedy in [True, False]:
-                                    _run_single_problem(path_to_instance, dive,
-                                                        greedy, cspy,
-                                                        pricing_strategy)
+        path_to_instance_type = INPUT_FOLDER / instance_type
+        for path_to_instance in path_to_instance_type.glob("*"):
+            if PERFORMANCE:
+                _run_single_problem(path_to_instance,
+                                    **PERFORMANCE_SOLVER_PARAMS[instance_type])
+            else:
+                for dive in [True, False]:
+                    for cspy in [True, False]:
+                        for pricing_strategy in [
+                                "BestPaths", "BestEdges1", "BestEdges2", "Exact"
+                        ]:
+                            for greedy in [True, False]:
+                                _run_single_problem(path_to_instance, dive,
+                                                    greedy, cspy,
+                                                    pricing_strategy)
 
 
 def run_parallel():
-
+    """Iterates through the instances using in parallel using CPU_COUNT.
+    """
     all_files = list([
-        os.path.join(root, f) for instance_type in INSTANCE_TYPES
-        for root, _, files in os.walk(os.path.join(INPUT_FOLDER, instance_type))
-        for f in files
+        path_to_instance for instance_type in INSTANCE_TYPES
[…]
