### 2020-07-12 Kuifje02/vrpy#42 @Kuifje02
> Anyone on a \ Windows\ machine pls

I am on Windows. When I run the tests (from \vrpy) I get :
```
_______________________________ ERROR collecting benchmarks/tests/test_cvrp_augerat.py ________________________________
ImportError while importing test module 'C:\Users\romain.montagne\Desktop\vrpy\benchmarks\tests\test_cvrp_augerat.py'.
Traceback:
benchmarks\tests\test_cvrp_augerat.py:1: in <module>
    from benchmarks.augerat_dataset import AugeratDataSet
E   ModuleNotFoundError: No module named 'benchmarks'
```

and similarly with the other tests. You think this is Windows specific?
