### 2020-07-12 Kuifje02/vrpy#42 @Kuifje02
> Still have to maybe move the ortools related tests to the examples folder maybe

Yes I agree! + this would satisfy the JOSS suggestion to assert the results of the examples (https://github.com/Kuifje02/vrpy/issues/53#issue-653177840)
