### 2020-07-13 Kuifje02/vrpy#42 @torressa
Not sure.. :thinking:  Are you doing
`python3 -m pytest benchmarks/`?
