### 2020-07-13 Kuifje02/vrpy#42 @torressa
Without all the sys paths you need to either do that or add an empty `__init__.py` file. (see [here](https://docs.pytest.org/en/stable/goodpractices.html))
