### 2020-07-13 Kuifje02/vrpy#42 @Kuifje02
> Not sure.. 🤔 Are you doing
> `python3 -m pytest benchmarks/`?

Actually I was just doing `pytest benchmarks/`. If I do `python3 -m pytest benchmarks/`, nothing happens (no error is raised, but no test is run either):
![image](https://user-images.githubusercontent.com/19840033/87315363-b8e91900-c524-11ea-834c-ec48068082cc.png)
