### 2020-07-13 Kuifje02/vrpy#42 @Kuifje02
Ok, I think I got it... for some reason, although  I am using python 3, I need to run `python -m pytest benchmarks` (and not `python3 ...`)
