### 2020-07-13 Kuifje02/vrpy#42 @torressa
Not sure, I've always used it cause `python` calls python2.7.
But I think it might be a an installation thing, installing python3.whatever after python2.whatever, so the name is already taken
