### 2020-07-22 4ad4298
separated check methods form VRP class
