+++ vrpy/check.py
@@ -1,0 +1,195 @@
+from networkx import DiGraph, shortest_path, NetworkXError, has_path
+import logging
+from time import time
+
+from .greedy import Greedy
+from .master_solve_pulp import MasterSolvePulp
+from .subproblem_lp import SubProblemLP
+from .subproblem_cspy import SubProblemCSPY
+from .subproblem_greedy import SubProblemGreedy
+from .clarke_wright import ClarkeWright, RoundTrip
+from .schedule import Schedule
+from .main import VehicleRoutingProblem
+
+
+def check_arguments(prob: VehicleRoutingProblem = None):
+    """Checks if arguments are consistent."""
+
+    #if prob is None raise instance
+    if prob == None:
+        raise ValueError("VRPy instance is None")
+
+    # If num_stops/load_capacity/duration are not integers
+    if prob.num_stops and (not isinstance(prob.num_stops, int)
+                           or prob.num_stops <= 0):
+        raise TypeError("Maximum number of stops must be positive integer.")
+    if prob.load_capacity:
+        for value in prob.load_capacity:
+            if not isinstance(value, int) or value <= 0:
+                raise TypeError("Load capacity must be positive integer.")
+    if prob.duration and (not isinstance(prob.duration, int)
+                          or prob.duration < 0):
+        raise TypeError("Maximum duration must be positive integer.")
+    strategies = [
+        "Exact",
+        "BestEdges1",
+        "BestEdges2",
+        "BestPaths",
+    ]
+    if prob._pricing_strategy not in strategies:
+        raise ValueError(
+            "Pricing strategy %s is not valid. Pick one among %s" %
+            (prob._pricing_strategy, strategies))
+    if prob.mixed_fleet:
+        if prob.load_capacity and prob.num_vehicles:
+            if not len(prob.load_capacity) == len(prob.num_vehicles):
+                raise ValueError(
+                    "Input arguments load_capacity and num_vehicles must have same dimension."
+                )
+        if prob.load_capacity and prob.fixed_cost:
+            if not len(prob.load_capacity) == len(prob.fixed_cost):
+                raise ValueError(
+                    "Input arguments load_capacity and fixed_cost must have same dimension."
+                )
+        if prob.num_vehicles and prob.fixed_cost:
+            if not len(prob.num_vehicles) == len(prob.fixed_cost):
+                raise ValueError(
+                    "Input arguments num_vehicles and fixed_cost must have same dimension."
+                )
+        for (i, j) in prob.G.edges():
+            if not isinstance(prob.G.edges[i, j]["cost"], list):
+                raise TypeError(
+                    "Cost attribute for edge (%s,%s) should be of type list")
+            if len(prob.G.edges[i, j]["cost"]) != prob._vehicle_types:
+                raise ValueError(
+                    "Cost attribute for edge (%s,%s) has dimension %s, should have dimension %s."
+                    % (i, j, len(
+                        prob.G.edges[i, j]["cost"]), prob._vehicle_types))
+
+
+def check_vrp(prob: VehicleRoutingProblem = None):
+    """Checks if graph is well defined."""
+
+    #if prob is None raise instance
+    if prob == None:
+        raise ValueError("VRPy instance is None")
+
+    # if G is not a DiGraph
+    if not isinstance(prob.G, DiGraph):
+        raise TypeError(
+            "Input graph must be of type networkx.classes.digraph.DiGraph.")
+    for v in ["Source", "Sink"]:
+        # If Source or Sink is missing
+        if v not in prob.G.nodes():
+            raise KeyError("Input graph requires Source and Sink nodes.")
+        # If Source has incoming edges
+        if len(list(prob.G.predecessors("Source"))) > 0:
+            raise NetworkXError("Source must have no incoming edges.")
[…]
