+++ vrpy/main.py
@@ -9,6 +9,7 @@ from .subproblem_cspy import SubProblemCSPY
 from .subproblem_greedy import SubProblemGreedy
 from .clarke_wright import ClarkeWright, RoundTrip
 from .schedule import Schedule
+from .check import check_arguments, check_consistency, check_feasibility, check_initial_routes, check_vrp
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
@@ -106,9 +107,8 @@ class VehicleRoutingProblem:
         self._best_value = None
         self._best_routes = []
         self._best_routes_as_graphs = []
-
         # Check if given inputs are consistent
-        self._check_vrp()
+        check_vrp(self)
 
         # Runtime for latest solve call
         self.comp_time = None
@@ -268,16 +268,16 @@ class VehicleRoutingProblem:
         else:
             self._vehicle_types = 1
         # Consistency checks
-        self._check_arguments()
+        check_arguments(self)
         # Setup fixed costs
         if self.fixed_cost:
             self._add_fixed_costs()
         # Setup default attributes if missing
         self._update_dummy_attributes()
         # Check options consistency
-        self._check_consistency()
+        check_consistency(self)
         # Check feasibility
-        self._check_feasibility()
+        check_feasibility(self)
         # Lock preassigned routes
         if self._preassignments:
             self._lock()
@@ -291,7 +291,7 @@ class VehicleRoutingProblem:
         """Initialization with feasible solution."""
         if self._initial_routes:
             # Initial solution is given as input
-            self._check_initial_routes()
+            check_initial_routes(self)
         else:
             # Initial solution is computed with Clarke & Wright (or round trips)
             self._get_initial_solution()
@@ -723,166 +723,6 @@ class VehicleRoutingProblem:
                 for u in self.G.predecessors("Sink"))
         # Keep a (deep) copy of the graph
         self._H = self.G.to_directed()
-
-    def _check_vrp(self):
-        """Checks if graph is well defined."""
-        # if G is not a DiGraph
-        if not isinstance(self.G, DiGraph):
-            raise TypeError(
-                "Input graph must be of type networkx.classes.digraph.DiGraph."
-            )
-        for v in ["Source", "Sink"]:
-            # If Source or Sink is missing
-            if v not in self.G.nodes():
-                raise KeyError("Input graph requires Source and Sink nodes.")
-            # If Source has incoming edges
-            if len(list(self.G.predecessors("Source"))) > 0:
-                raise NetworkXError("Source must have no incoming edges.")
-            # If Sink has outgoing edges
-            if len(list(self.G.successors("Sink"))) > 0:
-                raise NetworkXError("Sink must have no outgoing edges.")
-        # Roundtrips should always be possible
-        # Missing edges are added with a high cost
-        for v in self.G.nodes():
-            if v not in ["Source", "Sink"]:
-                if v not in self.G.successors("Source"):
-                    logger.warning("Source not connected to %s" % v)
-                    self.G.add_edge("Source", v, cost=1e10)
-                if v not in self.G.predecessors("Sink"):
-                    logger.warning("%s not connected to Sink" % v)
-                    self.G.add_edge(v, "Sink", cost=1e10)
-        # If graph is disconnected
-        if not has_path(self.G, "Source", "Sink"):
-            raise NetworkXError("Source and Sink are not connected.")
-        # If cost is missing
-        for (i, j) in self.G.edges():
-            if "cost" not in self.G.edges[i, j]:
-                raise KeyError("Edge (%s,%s) requires cost attribute" % (i, j))
-
-    def _check_arguments(self):
-        """Checks if arguments are consistent."""
-        # If num_stops/load_capacity/duration are not integers
[…]
