### 2020-07-23 36af950
Update run script, rename solomon instances
