### 2020-07-23 1da596a
Remove unused imports, rename solomon instance
