### 2020-07-23 5bdaf5c
changed inputs in check_functions
