+++ vrpy/check.py
@@ -2,33 +2,26 @@ from networkx import DiGraph, shortest_path, NetworkXError, has_path
 import logging
 from time import time
 
-from .greedy import Greedy
-from .master_solve_pulp import MasterSolvePulp
-from .subproblem_lp import SubProblemLP
-from .subproblem_cspy import SubProblemCSPY
-from .subproblem_greedy import SubProblemGreedy
-from .clarke_wright import ClarkeWright, RoundTrip
-from .schedule import Schedule
-from .main import VehicleRoutingProblem
 
-
-def check_arguments(prob: VehicleRoutingProblem = None):
+def check_arguments(num_stops: int = None,
+                    load_capacity: list = None,
+                    duration: int = None,
+                    pricing_strategy: str = None,
+                    mixed_fleet: bool = None,
+                    fixed_cost: bool = None,
+                    G: DiGraph = None,
+                    vehicle_types: int = None,
+                    num_vehicles: list = None):
     """Checks if arguments are consistent."""
 
-    #if prob is None raise instance
-    if prob == None:
-        raise ValueError("VRPy instance is None")
-
     # If num_stops/load_capacity/duration are not integers
-    if prob.num_stops and (not isinstance(prob.num_stops, int)
-                           or prob.num_stops <= 0):
+    if num_stops and (not isinstance(num_stops, int) or num_stops <= 0):
         raise TypeError("Maximum number of stops must be positive integer.")
-    if prob.load_capacity:
-        for value in prob.load_capacity:
+    if load_capacity:
+        for value in load_capacity:
             if not isinstance(value, int) or value <= 0:
                 raise TypeError("Load capacity must be positive integer.")
-    if prob.duration and (not isinstance(prob.duration, int)
-                          or prob.duration < 0):
+    if duration and (not isinstance(duration, int) or duration < 0):
         raise TypeError("Maximum duration must be positive integer.")
     strategies = [
         "Exact",
@@ -36,133 +29,122 @@ def check_arguments(prob: VehicleRoutingProblem = None):
         "BestEdges2",
         "BestPaths",
     ]
-    if prob._pricing_strategy not in strategies:
+    if pricing_strategy not in strategies:
         raise ValueError(
             "Pricing strategy %s is not valid. Pick one among %s" %
-            (prob._pricing_strategy, strategies))
-    if prob.mixed_fleet:
-        if prob.load_capacity and prob.num_vehicles:
-            if not len(prob.load_capacity) == len(prob.num_vehicles):
+            (pricing_strategy, strategies))
+    if mixed_fleet:
+        if load_capacity and num_vehicles:
+            if not len(load_capacity) == len(num_vehicles):
                 raise ValueError(
                     "Input arguments load_capacity and num_vehicles must have same dimension."
                 )
-        if prob.load_capacity and prob.fixed_cost:
-            if not len(prob.load_capacity) == len(prob.fixed_cost):
+        if load_capacity and fixed_cost:
+            if not len(load_capacity) == len(fixed_cost):
                 raise ValueError(
                     "Input arguments load_capacity and fixed_cost must have same dimension."
                 )
-        if prob.num_vehicles and prob.fixed_cost:
-            if not len(prob.num_vehicles) == len(prob.fixed_cost):
+        if num_vehicles and fixed_cost:
+            if not len(num_vehicles) == len(fixed_cost):
                 raise ValueError(
                     "Input arguments num_vehicles and fixed_cost must have same dimension."
                 )
-        for (i, j) in prob.G.edges():
-            if not isinstance(prob.G.edges[i, j]["cost"], list):
+        for (i, j) in G.edges():
+            if not isinstance(G.edges[i, j]["cost"], list):
[…]
