### 2020-07-23 d347c8e
'Refactored by Sourcery'
