### 2020-07-24 704d5d6
Add node tracking for diving heuristic
