### 2020-07-28 fac764c
HyperHeuristic and refactoring
