+++ vrpy/main.py
@@ -10,7 +10,7 @@ from .subproblem_greedy import SubProblemGreedy
 from .clarke_wright import ClarkeWright, RoundTrip
 from .schedule import Schedule
 from .check import check_arguments, check_consistency, check_feasibility, check_initial_routes, check_vrp
-#from .hyper_price_selection import halvardo_hyper
+from .hyperheuristic import HyperHeuristic
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
@@ -57,21 +57,20 @@ class VehicleRoutingProblem:
             True if heterogeneous fleet.
             Defaluts to False.
     """
-    def __init__(
-        self,
-        G,
-        num_stops=None,
-        load_capacity=None,
-        duration=None,
-        time_windows=False,
-        pickup_delivery=False,
-        distribution_collection=False,
-        drop_penalty=None,
-        fixed_cost=0,
-        num_vehicles=None,
-        periodic=None,
-        mixed_fleet=False,
-    ):
+    def __init__(self,
+                 G,
+                 num_stops=None,
+                 load_capacity=None,
+                 duration=None,
+                 time_windows=False,
+                 pickup_delivery=False,
+                 distribution_collection=False,
+                 drop_penalty=None,
+                 fixed_cost=0,
+                 num_vehicles=None,
+                 periodic=None,
+                 mixed_fleet=False,
+                 use_hyper_heuristic=True):
         self.G = G
         # VRP options/constraints
         self.num_stops = num_stops
@@ -87,6 +86,7 @@ class VehicleRoutingProblem:
         self.mixed_fleet = mixed_fleet
         # Parameters for solving
         self.masterproblem = None
+
         self.routes = []
         self._solver = None
         self._time_limit = None
@@ -114,7 +114,9 @@ class VehicleRoutingProblem:
         # Runtime for latest solve call
         self.comp_time = None
 
-        # Initialise the HyperHeuristic solver
+        # Initialise the HyperHeuristic solver, for pricing strategies
+        self.hyper_heuristic = HyperHeuristic(poolsize=4)
+        self.use_hyper_heuristic = use_hyper_heuristic
 
     def solve(self,
               initial_routes=None,
@@ -126,7 +128,8 @@ class VehicleRoutingProblem:
               solver="cbc",
               dive=False,
               greedy=False,
-              compute_runtime=False):
+              compute_runtime=False,
+              use_hyper_heuristic=True):
         """Iteratively generates columns with negative reduced cost and solves as MIP.
 
         Args:
@@ -173,10 +176,16 @@ class VehicleRoutingProblem:
             compute_runtime (bool, optional):
                 True if solve runtime is to be returned
                 Defaults to False. 
+            use_hyper_heuristic (bool, optional):
+                True if hyper_heuristic is to be employed
+                Defaults to True 
 
         Returns:
             float: Optimal solution of MIP based on generated columns
         """
+
+        self.use_hyper_heuristic = use_hyper_heuristic
+
         # compute run time
         if compute_runtime:
             starttime = time()
@@ -324,11 +333,65 @@ class VehicleRoutingProblem:
             solver,
         )
 
+    def _solve_subproblem_with_heuristic(self,
+                                         pricing_strategy=None,
+                                         vehicle=None,
+                                         duals=None):
+
+        if not self._more_routes and self._pricing_strategy == "BestPaths":
+            for k_shortest_paths in [3, 5, 7, 9]:
+                subproblem = self._def_subproblem(
+                    duals,
+                    vehicle,
+                    "BestPaths",
+                    k_shortest_paths,
+                )
+
+                self.routes, self._more_routes = subproblem.solve(
+                    self._get_time_remaining())
+                if self._more_routes:
[…]
