### 2020-07-31 bab5d2b
hyperheurstic and main changes, along with run.py changes
