+++ vrpy/hyperheuristic.py
@@ -1,14 +1,33 @@
 from random import uniform, choice
 from math import exp, log, sqrt
+from time import time
+import json
 #from .main import VehicleRoutingProblem
 """
 TODO: 
-    - Bruke reduced cost på en måte! (Multiobjective optimisiation)
-    - Gi muligheten for å velge en spesifikk heurestikk
+    - Bruke reduced cost på en måte! (Multiobjective optimisiation) GO R
+    - Gi muligheten for å velge en spesifikk heurestikk R, kan gjøre bedre
     - Straffe heurestikken hvis det ikke er forbedring, ved å øke n[i]
-    - Change the move acceptance
     - Det sto noe i den andre om rank based move acceptance, dette bør sjekkes ut fordi vi bryr oss om hvorvidt nye kolonner er funnet, men ikke om verdier!
 """
+# tre oppgaver: learning R for now, update data structures R , and change the punishment criterium Lets go!
+
+# påstand4 (big step), hva med å ordne heurestikkene i en prioritetskø?
+
+#Oppgaver
+# Finn en måte å sammenlikne resultatet av metaheurestikken på de ulike instancene!
+# Finn ut best måte å teste performance av metaheurestikken
+# Kjøre performance profileren på en rekke instances, basically!
+# ting jeg setter: move_acceptance
+# ting som jeg vil ha med.
+# reject list, n_list,
+# Sammenlikne Kjøretid i move_acceptance criteria
+# ta stilling til om jeg vil kjøre på
+
+# SLITEN:
+#navnendringer:
+"hyperparams"
+#sette inn types i metodene
 
 
 class HyperHeuristic:
@@ -47,7 +66,11 @@ class HyperHeuristic:
         current_heuristic (int):
             the currently applied heuristic
     """
-    def __init__(self, poolsize=4, start_heur=0, scaling_factor=0.01):
+    def __init__(self,
+                 poolsize: int = 4,
+                 start_heur: int = 0,
+                 scaling_factor: float = 0.01,
+                 offline_learning: bool = False):
 
         #params
         self.scaling_factor = scaling_factor
@@ -69,6 +92,9 @@ class HyperHeuristic:
         self.q = [0] * poolsize
         self.r = [0] * poolsize
         self.n = [0] * poolsize
+        self.d = 0
+
+        #Consider setting to zero
         self.heuristic_points = [self.inf] * poolsize
         self.improvements = {}
         for element in range(poolsize):
@@ -76,17 +102,39 @@ class HyperHeuristic:
 
         #consider
         self.current_heuristic = start_heur
+        self.loaded_parameters = {}
 
-    def set_current_objective(self, objective=None):
+        #total time
+        self.timestart = None
+        self.timeend = None
+        self.time_windows = None
+        self.pos_reduced_cost = None
+        self.reject_list = [0, 0, 0, 0]
+        self.total_n = 0
+        self.start_computing_average = 3
+        self.last_runtime = None
+        self.average_runtime = 0
+
+    def set_current_objective(self, objective: float = None):
         """
         sets current_objective_value
         """
         self.current_objective_value = objective
 
-    def pick_heurestic(self):
+    def pick_heurestic(self, heuristic: int = None):
         """
         Sets the chosen heuristic based on heuristic points
         """
+
+        #force a particular heuristic
+        if not heuristic == None:
+            return heuristic
+
+        #before the number of iterations is high enough
+        if self.total_n < self.start_computing_average:
+            return choice(self.pool)
+
+        #choose according to MAB
         maxval = max(self.heuristic_points)
         best_heuristics = [
             i for i, j in enumerate(self.heuristic_points) if j == maxval
@@ -98,7 +146,42 @@ class HyperHeuristic:
[…]
