+++ vrpy/subproblem_cspy.py
@@ -4,8 +4,9 @@ from math import floor
 from numpy import array, zeros
 from networkx import DiGraph, add_path
 
-from cspy import BiDirectional, GreedyElim  # Tabu
-from .subproblem import SubProblemBase
+from cspy import BiDirectional
+
+from vrpy.subproblem import SubProblemBase
 
 logger = logging.getLogger(__name__)
 
@@ -22,8 +23,8 @@ class SubProblemCSPY(SubProblemBase):
         """Initializes resources."""
         # Pass arguments to base
         super(SubProblemCSPY, self).__init__(*args)
+        self.exact = exact
         # Resource names
-        self.alg = None
         self.resources = [
             "stops/mono",
             "load",
@@ -32,7 +33,6 @@ class SubProblemCSPY(SubProblemBase):
             "collect",
             "deliver",
         ]
-        self.exact = exact
         # Set number of resources as attribute of graph
         self.sub_G.graph["n_res"] = len(self.resources)
         # Default lower and upper bounds
@@ -79,35 +79,39 @@ class SubProblemCSPY(SubProblemBase):
 
         while True:
             if self.exact:
-                logger.debug("solving with bidirectional")
-                self.alg = BiDirectional(
+                alg = BiDirectional(
                     self.sub_G,
                     self.max_res,
                     self.min_res,
                     direction="both",
                     method="generated",
+                    time_limit=time_limit,
                     REF_forward=self.get_REF("forward"),
                     REF_backward=self.get_REF("backward"),
                     REF_join=self.get_REF("join"),
                 )
             else:
-                logger.debug("solving with greedyelim")
-                self.alg = GreedyElim(
+                alg = BiDirectional(
                     self.sub_G,
                     self.max_res,
                     self.min_res,
-                    REF=self.get_REF("forward"),
-                    max_depth=40,
+                    direction="both",
+                    method="generated",
+                    time_limit=time_limit,
+                    threshold=-1,
+                    REF_forward=self.get_REF("forward"),
+                    REF_backward=self.get_REF("backward"),
+                    REF_join=self.get_REF("join"),
                 )
-            self.alg.run()
+            alg.run()
             logger.debug("subproblem")
-            logger.debug("cost = %s" % self.alg.total_cost)
-            logger.debug("resources = %s" % self.alg.consumed_resources)
-            if self.alg.total_cost < -(10**-3):
+            logger.debug("cost = %s" % alg.total_cost)
+            logger.debug("resources = %s" % alg.consumed_resources)
+            if alg.total_cost < -(1e-3):
                 more_routes = True
-                self.add_new_route()
-                logger.debug("new route %s" % self.alg.path)
-                logger.debug("reduced cost = %s" % self.alg.total_cost)
+                self.add_new_route(alg.path)
+                logger.debug("new route %s" % alg.path)
+                logger.debug("reduced cost = %s" % alg.total_cost)
                 logger.debug("real cost = %s" % self.total_cost)
                 break
             # If not already solved exactly
@@ -137,22 +141,20 @@ class SubProblemCSPY(SubProblemBase):
             # Time windows feasibility
             self.max_res[3] = 0
             # Maximum feasible arrival time
-            self.T = max(
-                self.sub_G.nodes[v]["upper"]
-                + self.sub_G.nodes[v]["service_time"]
-                + self.sub_G.edges[v, "Sink"]["time"]
-                for v in self.sub_G.predecessors("Sink")
-            )
+            self.T = max(self.sub_G.nodes[v]["upper"] +
+                         self.sub_G.nodes[v]["service_time"] +
[…]
