+++ vrpy/hyperheuristic.py
@@ -3,31 +3,27 @@ from math import exp, log, sqrt
 from time import time
 import json
 #from .main import VehicleRoutingProblem
+import logging
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
 """
 TODO: 
-    - Bruke reduced cost på en måte! (Multiobjective optimisiation) GO R
-    - Gi muligheten for å velge en spesifikk heurestikk R, kan gjøre bedre
-    - Straffe heurestikken hvis det ikke er forbedring, ved å øke n[i]
-    - Det sto noe i den andre om rank based move acceptance, dette bør sjekkes ut fordi vi bryr oss om hvorvidt nye kolonner er funnet, men ikke om verdier!
-"""
-# tre oppgaver: learning R for now, update data structures R , and change the punishment criterium Lets go!
-
-# påstand4 (big step), hva med å ordne heurestikkene i en prioritetskø?
-
-#Oppgaver
-# Finn en måte å sammenlikne resultatet av metaheurestikken på de ulike instancene!
-# Finn ut best måte å teste performance av metaheurestikken
+    #Oppgaver
 # Kjøre performance profileren på en rekke instances, basically!
-# ting jeg setter: move_acceptance
-# ting som jeg vil ha med.
-# reject list, n_list,
-# Sammenlikne Kjøretid i move_acceptance criteria
-# ta stilling til om jeg vil kjøre på
+# Fikse i 1) main 2) hyper 3) csv_writer
+# Finne noen gode instancer å se på
+# Resette average runtime, vurdere -> raskere enn previous search markov chain
+
+# Så kan vi se mer på move_acceptance
 
 # SLITEN:
-#navnendringer:
-"hyperparams"
-#sette inn types i metodene
+# sortere variablene
+# åpne for muligheten å velge en spesifikk heurestikk
+    - Det sto noe i den andre om rank based move acceptance, dette bør sjekkes ut fordi vi bryr oss om hvorvidt nye kolonner er funnet, men ikke om verdier! (kanskje)
+# ordne modulo greiene 
+
+"""
 
 
 class HyperHeuristic:
@@ -69,16 +65,18 @@ class HyperHeuristic:
     def __init__(self,
                  poolsize: int = 4,
                  start_heur: int = 0,
-                 scaling_factor: float = 0.01,
+                 scaling_factor: float = 0.5,
                  offline_learning: bool = False):
 
         #params
         self.scaling_factor = scaling_factor
+        self.theta = 1
         #self.decay_factor = None
 
         #self.time_limit = None
         self.current_objective_value = None
         self.new_objective_value = None
+        self.pos_reduced_cost = None  #ikke egentlig positive reduce cost at this point
 
         self.inf = 1E10
 
@@ -92,7 +90,11 @@ class HyperHeuristic:
         self.q = [0] * poolsize
         self.r = [0] * poolsize
         self.n = [0] * poolsize
+        self.exp_list = [0] * poolsize
         self.d = 0
+        self.d_max = 0
+        self.total_n = 0
+        self.average_runtime = 0
 
         #Consider setting to zero
         self.heuristic_points = [self.inf] * poolsize
@@ -107,13 +109,10 @@ class HyperHeuristic:
         #total time
         self.timestart = None
         self.timeend = None
-        self.time_windows = None
-        self.pos_reduced_cost = None
-        self.reject_list = [0, 0, 0, 0]
-        self.total_n = 0
-        self.start_computing_average = 3
         self.last_runtime = None
-        self.average_runtime = 0
+        self.time_windows = None
+        self.start_computing_average = 1
+        self.reject_list = [0, 0, 0, 0]
 
[…]
