### 2020-08-04 dc144bb
update main, heuristic, checks and run
