### 2020-08-05 ad2c9f9
Disable pulp logs in subproblem #53
