### 2020-08-05 17ccf94
update test_solomon and subproblem_lp
