### 2020-08-07 Kuifje02/vrpy#44 @Kuifje02
@torressa what do you think of this version of the paper ?
- Maybe a word on the hyperheuristic ?
- Maybe performances on Augerat data sets have improved ?
