### 2020-08-07 Kuifje02/vrpy#44 @torressa
It's looking really good! I've just updated the results in `performance_profiles` with some new ones.
Also changed the notebook and csv files so they can be rendered on github (except for the res column in the vrpy file because I didn't have that)

@Halvaros maybe you can write something about the hyper-heuristic in the paper too?
I guess you'll be co-author as well :)
