### 2020-08-10 f6f0b03
Change master_solve_pulp, refactor in main, termination criterium in solve
