+++ vrpy/main.py
@@ -1,7 +1,7 @@
 import logging
 from time import time
 
-from networkx import DiGraph, shortest_path
+from networkx import DiGraph, shortest_path  #draw_networkx
 
 from vrpy.greedy import Greedy
 from vrpy.master_solve_pulp import MasterSolvePulp
@@ -121,9 +121,12 @@ class VehicleRoutingProblem:
         self.end_time = None
 
         # Initialise the HyperHeuristic solver, for pricing strategies
-        self.hyper_heuristic = HyperHeuristic(poolsize=4)
+        self.hyper_heuristic = HyperHeuristic()
         self.use_hyper_heuristic = use_hyper_heuristic
-        self.resolve_count = None
+        self.produced_column = None
+
+        #number of iterations before exact is tried
+        self.do_exact = 30
 
     def solve(self,
               initial_routes=None,
@@ -338,6 +341,92 @@ class VehicleRoutingProblem:
             solver,
         )
 
+    def _attempt_solve_best_paths(self,
+                                  pricing_strategy=None,
+                                  vehicle=None,
+                                  duals=None):
+        produced_column = False
+        for k_shortest_paths in [3, 5, 7, 9]:
+            subproblem = self._def_subproblem(
+                duals,
+                vehicle,
+                "BestPaths",
+                k_shortest_paths,
+            )
+            logger.info("k_shortest paths %s" % k_shortest_paths)
+            self.routes, self._more_routes = subproblem.solve(
+                self._get_time_remaining())
+            produced_column = self._more_routes
+            if self._more_routes:
+                break
+        else:
+            self._more_routes = True
+            logger.info("Heuristic failed,s et to exact")
+        return produced_column
+
+    def _attempt_solve_BestEdges1(self,
+                                  pricing_strategy=None,
+                                  vehicle=None,
+                                  duals=None):
+        produced_column = False
+        for alpha in [0.3, 0.5, 0.7, 0.9]:
+            subproblem = self._def_subproblem(
+                duals,
+                vehicle,
+                "BestEdges1",
+                alpha,
+            )
+            self.routes, self._more_routes = subproblem.solve(
+                self._get_time_remaining(),
+                # exact=False,
+            )
+            produced_column = self._more_routes
+            logger.info("alpha paths %s" % alpha)
+            if self._more_routes:
+                break
+        else:
+            self._more_routes = True
+            logger.info("Heuristic failed,s et to exact")
+        return produced_column
+
+    def _attempt_solve_BestEdges2(self,
+                                  pricing_strategy=None,
+                                  vehicle=None,
+                                  duals=None):
+        produced_column = False
+        for ratio in [0.1, 0.2, 0.3]:
+            subproblem = self._def_subproblem(
+                duals,
+                vehicle,
+                "BestEdges2",
+                ratio,
+            )
+            self.routes, self._more_routes = subproblem.solve(
+                self._get_time_remaining(),
+                # exact=False,
+            )
+            produced_column = self._more_routes
+            logger.info("ratio %s" % ratio)
+            if self._more_routes:
+                break
+        else:
+            self._more_routes = True
+            logger.info("Heuristic failed,s et to exact")
+        return produced_column
+
+    def _attempt_solve_Exact(self,
+                             pricing_strategy=None,
+                             vehicle=None,
+                             duals=None):
+        logger.info("Run exact")
+        subproblem = self._def_subproblem(duals, vehicle)
+        self.routes, self._more_routes = subproblem.solve(
+            self._get_time_remaining(),
+            # exact=False,
+        )
+        produced_column = self._more_routes
+        return produced_column
+
[…]
