### 2020-08-11 #2408 @torressa
@kakiac @bstabler Quick question, is it possible to add a new co-author at this stage?

We'd like to include @Halvaros as he's done a lot of work over the summer including some new hyper-heuristic selection functionality that picks the pricing heuristics on the fly.
