### 2020-08-11 #2408 @Kuifje02
@kakiac @bstabler While we are at it, another question. We have taken into account all of the comments for the paper, but the resulting paper is longer than the 1000 word limit (~ 1300 words). Is the 1000 word limit strict ?
