### 2020-08-13 29b65f8
New option hyper, added visualisation
