+++ vrpy/main.py
@@ -1,5 +1,6 @@
 import logging
 from time import time
+from pathlib import Path
 
 from networkx import DiGraph, shortest_path  #draw_networkx
 
@@ -13,6 +14,7 @@ from vrpy.schedule import Schedule
 from vrpy.checks import (check_arguments, check_consistency, check_feasibility,
                          check_initial_routes, check_vrp)
 from .hyperheuristic import HyperHeuristic
+from csv import DictWriter
 
 logger = logging.getLogger(__name__)
 
@@ -126,7 +128,7 @@ class VehicleRoutingProblem:
         self.produced_column = None
 
         #number of iterations before exact is tried
-        self.do_exact = 30
+        self.do_exact = 1000
 
     def solve(self,
               initial_routes=None,
@@ -353,7 +355,6 @@ class VehicleRoutingProblem:
                 "BestPaths",
                 k_shortest_paths,
             )
-            logger.info("k_shortest paths %s" % k_shortest_paths)
             self.routes, self._more_routes = subproblem.solve(
                 self._get_time_remaining())
             produced_column = self._more_routes
@@ -361,7 +362,6 @@ class VehicleRoutingProblem:
                 break
         else:
             self._more_routes = True
-            logger.info("Heuristic failed,s et to exact")
         return produced_column
 
     def _attempt_solve_BestEdges1(self,
@@ -381,12 +381,10 @@ class VehicleRoutingProblem:
                 # exact=False,
             )
             produced_column = self._more_routes
-            logger.info("alpha paths %s" % alpha)
             if self._more_routes:
                 break
         else:
             self._more_routes = True
-            logger.info("Heuristic failed,s et to exact")
         return produced_column
 
     def _attempt_solve_BestEdges2(self,
@@ -406,12 +404,11 @@ class VehicleRoutingProblem:
                 # exact=False,
             )
             produced_column = self._more_routes
-            logger.info("ratio %s" % ratio)
             if self._more_routes:
                 break
         else:
             self._more_routes = True
-            logger.info("Heuristic failed,s et to exact")
+
         return produced_column
 
     def _attempt_solve_Exact(self,
@@ -438,17 +435,7 @@ class VehicleRoutingProblem:
         """TODO: change to if (self._no_improvement >= K and not self._more_routes and self.exact) or pricing_strategy == "Exact":"""
         produced_column = False
         if self._pricing_strategy == "Hyper":
-
-            if pricing_strategy == "Exact" or self._no_improvement > self.do_exact:
-                self._no_improvement = 0  #Fudge solution
-                print("Did exact")
-                self.hyper_heuristic.n_exact += 1
-                produced_column = self._attempt_solve_Exact(
-                    pricing_strategy=pricing_strategy,
-                    vehicle=vehicle,
-                    duals=duals)
-
-            elif pricing_strategy == "BestPaths":
+            if pricing_strategy == "BestPaths":
                 produced_column = self._attempt_solve_best_paths(
                     pricing_strategy=pricing_strategy,
                     vehicle=vehicle,
@@ -462,6 +449,15 @@ class VehicleRoutingProblem:
 
             elif pricing_strategy == "BestEdges2":
                 produced_column = self._attempt_solve_BestEdges2(
+                    pricing_strategy=pricing_strategy,
+                    vehicle=vehicle,
+                    duals=duals)
+
+            elif pricing_strategy == "Exact":
+                #self._no_improvement = 0  #Fudge solution
+                print("Did exact")
+                self.hyper_heuristic.n_exact += 1
+                produced_column = self._attempt_solve_Exact(
                     pricing_strategy=pricing_strategy,
                     vehicle=vehicle,
                     duals=duals)
[…]
