+++ vrpy/restricted_master_heuristics.py
@@ -1,3 +1,11 @@
+"""
+File to hold different restricted master heuristics.
+Will possible move into a single class after implemnting the algorithm
+in issue #60 if there is enough overlap.
+
+Currently implemented is
+- Diving heuristic
+"""
 import logging
 from typing import List, Optional
 
@@ -6,7 +14,13 @@ import pulp
 logger = logging.getLogger(__name__)
 
 
-class PricingHeuristics:
+class DivingHeuristic:
+    """Implements diving algorithm with Limited Discrepancy Search
+    Parameters as suggested by the authors. This only fixes one column.
+    `Sadykov et al. (2019)`_.
+
+    .. _Sadykov et al. (2019): https://pubsonline.informs.org/doi/abs/10.1287/ijoc.2018.0822
+    """
 
     def __init__(self, max_depth: int = 3, max_discrepancy: int = 1):
         self.max_depth = max_depth
@@ -16,12 +30,6 @@ class PricingHeuristics:
         self.tabu_list = []
 
     def run_dive(self, prob):
-        """Implements diving algorithm with Limited Discrepancy Search
-        Parameters as suggested by the authors. This only fixes one column.
-        `Sadykov et al. (2019)`_.
-
-        .. _Sadykov et al. (2019): https://pubsonline.informs.org/doi/abs/10.1287/ijoc.2018.0822
-        """
         tabu_list = []
         relax = prob.deepcopy()
         # Init current_node
