+++ benchmarks/tests/test_cvrptw_solomon.py
@@ -1,13 +1,10 @@
-""" from sys import path
-
-path.append('/mnt/c/Users/Halvardo/Documents/code/vrpy') """
+from benchmarks.solomon_dataset import SolomonDataSet
 
 from vrpy import VehicleRoutingProblem
 
-from benchmarks.solomon_dataset import SolomonDataSet
-
 
 class TestsSolomon:
+
     def setup(self):
         """
         Solomon instance c101, 25 first nodes only including depot
