### 2020-08-14 67ebee0
refactor hyper update  main, readme
