+++ vrpy/hyperheuristic.py
@@ -14,59 +14,52 @@ class HyperHeuristic:
     """
     HyperHeuristic class manages the high-level heuristic strategies
     Args: 
+        start_heur (int): 
+            set the initialisation heuristic. Defaults to BestPaths
         scaling_factor (float): 
-            parameter deciding the balance between exploration and explotation
-        time_limit (float): 
-            total time limit for the algorithm
-        current_objective_value (float): 
-            the objective value for previous iteration
-        new_objective_value (float): 
-            the objective value for the current iteration
-        inf (float): 
-            very large number 
-        iteration_counter (int): 
-            counts the number of iterations
-        initialisation (bool): 
-            initialisation parameter 
-        heur_names (list, str): 
-            names of heurestic strategies 
-        pool (list, int: 
-            the pool of heurestics, indexed by integers
-        q (list, floats): 
-            list of heuristic quality (rank)
-        r (list, floats): 
-            list of average improvements
-        n (list, int): 
-            list of iteration counters
-        heuristic_points (list, floats): 
-            list of points for each heurestic at the current timepoint
-        improvements (dict, dict, floats):
-            a dictionary with keys indexed by heuristics with values equal to dictionaries with rewards
-        current_heuristic (int):
-            the currently applied heuristic
+            set the evolution vs exploration parameter. Defaults to 0.5
+        performance_measure (str):
+            Sets performance measure. Defaults to "Weighted average".
+        acceptance_type(str):
+            Set move acceptance. Defaults to "Accept all"
+        step (float, optional): 
+            Adaptive step. Defaults to 0.02
+        start_computing_average (int): 
+            Iteration where average starts being computed. Defaults to 1. 
+
     """
     def __init__(self,
-                 poolsize: int = 4,
                  start_heur: int = 0,
                  scaling_factor: float = 0.5,
-                 offline_learning: bool = False,
                  performance_measure="Weighted average",
-                 acceptance_type="Accept all"):
+                 acceptance_type="Accept all",
+                 step=0.02,
+                 start_computing_average=1):
 
         #params
         self.scaling_factor = scaling_factor
-        self.theta = 1
-        #self.decay_factor = None
+        self.current_heuristic = start_heur
+        self.start_computing_average = start_computing_average
+        self.step = step
 
-        #self.time_limit = None
+        #Settings
+        self.performance_measure = performance_measure
+        self.acceptance_type = acceptance_type
+
+        if self.performance_measure == "Weighted average":
+            poolsize = 4
+        elif self.performance_measure == "Relative improvement":
+            poolsize = 3
+        else:
+            poolsize = 4
+
+        #adaptive CF function
+        self.theta = 1
         self.current_objective_value = None
         self.new_objective_value = None
         self.produced_column = None
-
         self.inf = 1E10
 
-        #consider
-        #self.iteration_counter = 0  # maybe one for each heuristic?
         self.initialisation = True
 
         #high level data
@@ -78,16 +71,16 @@ class HyperHeuristic:
         self.q = [0] * poolsize
         self.r = [0] * poolsize
         self.n = [0] * poolsize
+        self.heuristic_points = [self.inf] * poolsize
         self.added_columns = [0] * poolsize
         self.exp_list = [0] * poolsize
         self.d = 0
         self.d_max = 0
         self.total_n = 0
         self.average_runtime = 0
-        self.step = 0.02
         self.n_exact = 0
 
-        #weighed average approach
+        #weighed average performance measure
         self.runtime_dist = []
         self.objective_decrease_list = [0] * poolsize
         self.norm_objective_decrease_list = [0] * poolsize
@@ -95,23 +88,22 @@ class HyperHeuristic:
         self.norm_runtime_list = [0] * poolsize
[…]
