+++ vrpy/hyperheuristic.py
@@ -156,7 +156,7 @@ class HyperHeuristic:
         self.timeend = time()
         self.last_runtime = self.timeend - self.timestart
 
-    def _current_performance_relimp(self):
+    def _current_performance_relimp(self, produced_column: bool = False):
         #   time measure
         if self.total_n > self.start_computing_average + 1:
             self.d = max((self.average_runtime - self.last_runtime) /
@@ -228,7 +228,7 @@ class HyperHeuristic:
         self.obj_has_decreased = self.current_objective_value - new_objective_value > 0
 
         if self.performance_measure == "Relative improvement":
-            self._current_performance_relavr()
+            self._current_performance_relimp(produced_column=produced_column)
 
         elif self.performance_measure == "Weighted average":
             self._current_performance_wgtavr(
@@ -300,79 +300,76 @@ class HyperHeuristic:
         """Updates params for relative improvements performance measure"""
 
         if self.total_n > self.start_computing_average:
-                reduced_n = (self.total_n - self.start_computing_average) % 10
-                if reduced_n == 0:
-                    self.average_runtime = self.last_runtime
-                else:
-                    self.average_runtime = self.average_runtime * (
-                        reduced_n -
-                        1) / (reduced_n) + 1 / (reduced_n) * self.last_runtime
+            i = self.current_heuristic
+            reduced_n = (self.total_n - self.start_computing_average) % 10
+            if reduced_n == 0:
+                self.average_runtime = self.last_runtime
+            else:
+                self.average_runtime = self.average_runtime * (
+                    reduced_n -
+                    1) / (reduced_n) + 1 / (reduced_n) * self.last_runtime
 
-                #store old values
-                old_q = self.q[i]
-                old_n = self.n[i] - 1
+            #store old values
+            old_q = self.q[i]
+            old_n = self.n[i] - 1
 
-                frozen = (old_q == 0 and old_n > 3)
+            frozen = (old_q == 0 and old_n > 3)
 
-                #average of improvements
-                self.r[i] = self.r[i] * old_n / (old_n + 1) + 1 / (
-                    old_n + 1) * self.reward(self.d, lower_bound=frozen)
+            #average of improvements
+            self.r[i] = self.r[i] * old_n / (old_n + 1) + 1 / (
+                old_n + 1) * self.reward(self.d, lower_bound=frozen)
 
-                self.q[i] = (old_q + self.r[i]) / (old_n + 1)
+            self.q[i] = (old_q + self.r[i]) / (old_n + 1)
 
-                #compute heuristic points MAB-style
-                for j in range(len(self.pool)):
-                    if not self.n[j] == 0:
-                        self.exp_list[j] = sqrt(2 * log(sum(self.n)) /
-                                                self.n[j])
-                        #OPTION1: theta f1 +  (1 - theta) f2
-                        """ self.heuristic_points[
+            #compute heuristic points MAB-style
+            for j in range(len(self.pool)):
+                if not self.n[j] == 0:
+                    self.exp_list[j] = sqrt(2 * log(sum(self.n)) / self.n[j])
+                    #OPTION1: theta f1 +  (1 - theta) f2
+                    """ self.heuristic_points[
                             j] = self.theta * self.q[j] + self.scaling_factor(
                                 1 - self.theta) * self.exp_list[
                                     j]  # consider adaptive selection """
-                        #OPTION2: thetaf1 + f2
-                        self.heuristic_points[j] = self.theta * self.q[
-                            j] + self.scaling_factor * self.exp_list[
-                                j]  # consider adaptive selection
+                    #OPTION2: thetaf1 + f2
+                    self.heuristic_points[j] = self.theta * self.q[
+                        j] + self.scaling_factor * self.exp_list[
[…]
