+++ benchmarks/augerat_dataset.py
@@ -58,7 +58,7 @@ class AugeratDataSet:
                 if i == 1:
                     best = line.split()[-1][:-1]
                     self.best_known_solution = int(best)
-                if i == 5:
+                elif i == 5:
                     self.max_load = int(line.split()[2])
         fp.close()
         # Create network and store name + capacity
@@ -103,12 +103,11 @@ class AugeratDataSet:
         for u in self.G.nodes():
             if u != "Sink":
                 for v in self.G.nodes():
-                    if v != "Source":
-                        if u != v:
-                            self.G.add_edge(u,
-                                            v,
-                                            cost=round(distance(self.G, u, v),
-                                                       1))
+                    if v != "Source" and u != v:
+                        self.G.add_edge(u,
+                                        v,
+                                        cost=round(distance(self.G, u, v),
+                                                   1))
 
         # relabel
         before = [v for v in self.G.nodes() if v not in ["Source", "Sink"]]
