+++ examples/benchmarks/not committed/mdvrp_cordeau.py
@@ -50,7 +50,7 @@ class DataSet:
                         self.n_vertices = min(self.n_customers, n_vertices)
                     else:
                         self.n_vertices = self.n_customers
-                if i == 2:
+                elif i == 2:
                     self.max_load = int(line.split()[1])
         fp.close()
 
@@ -116,10 +116,7 @@ class DataSet:
 
     def solve(self, initial_routes=None, cspy=False):
         """Instantiates instance as VRP and solves."""
-        if cspy:
-            self.G.graph["subproblem"] = "cspy"
-        else:
-            self.G.graph["subproblem"] = "lp"
+        self.G.graph["subproblem"] = "cspy" if cspy else "lp"
         print(self.G.graph["name"], self.G.graph["subproblem"])
         print("===========")
         prob = VehicleRoutingProblem(self.G, load_capacity=self.max_load,)
