+++ tests/test_consistency.py
@@ -1,6 +1,8 @@
 from networkx import DiGraph, Graph, NetworkXError
 import pytest
+import sys
 
+sys.path.append("../")
 from vrpy.main import VehicleRoutingProblem
 
 #####################
@@ -68,10 +70,9 @@ def test_mixed_fleet_consistency():
         prob.solve()
     G.edges["Source", "Sink"]["cost"] = [1, 2]
     with pytest.raises(ValueError):
-        prob = VehicleRoutingProblem(G,
-                                     mixed_fleet=True,
-                                     load_capacity=[2, 4],
-                                     fixed_cost=[4])
+        prob = VehicleRoutingProblem(
+            G, mixed_fleet=True, load_capacity=[2, 4], fixed_cost=[4]
+        )
         prob.solve()
 
 
