+++ tests/pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
 python_files = test_*.py
-testpaths = tests/
+#testpaths = tests/
 filterwarnings =
     ignore::DeprecationWarning
