+++ vrpy/subproblem_lp.py
@@ -20,9 +20,7 @@ class SubProblemLP(SubProblemBase):
         # create problem
         self.prob = pulp.LpProblem("SubProblem", pulp.LpMinimize)
         # flow variables
-        self.x = pulp.LpVariable.dicts("x",
-                                       self.sub_G.edges(),
-                                       cat=pulp.LpBinary)
+        self.x = pulp.LpVariable.dicts("x", self.sub_G.edges(), cat=pulp.LpBinary)
         self.solver = solver
 
     # @profile
@@ -41,10 +39,10 @@ class SubProblemLP(SubProblemBase):
         logger.debug("Solving subproblem using LP")
         logger.debug("Status: %s" % pulp.LpStatus[self.prob.status])
         logger.debug("Objective %s" % pulp.value(self.prob.objective))
-        if (pulp.value(self.prob.objective) is not None and
-                pulp.value(self.prob.objective) < -(10**-3)) or (
-                    exact == False and
-                    pulp.LpStatus[self.prob.status] in ["Optimal", ""]):
+        if (
+            pulp.value(self.prob.objective) is not None
+            and pulp.value(self.prob.objective) < -(10 ** -3)
+        ) or (exact == False and pulp.LpStatus[self.prob.status] in ["Optimal", ""]):
             more_routes = True
             self._add_new_route()
         else:
@@ -68,10 +66,10 @@ class SubProblemLP(SubProblemBase):
         new_route.graph["vehicle_type"] = self.vehicle_type
         self.routes.append(new_route)
 
-        logger.debug("new route %s %s" %
-                     (route_id, shortest_path(new_route, "Source", "Sink")))
-        logger.debug("new route reduced cost %s" %
-                     pulp.value(self.prob.objective))
+        logger.debug(
+            "new route %s %s" % (route_id, shortest_path(new_route, "Source", "Sink"))
+        )
+        logger.debug("new route reduced cost %s" % pulp.value(self.prob.objective))
         logger.debug("new route cost = %s" % self.total_cost)
         # for (i, j) in new_route.edges():
         #    print(i, j, self.sub_G.edges[i, j])
@@ -82,42 +80,41 @@ class SubProblemLP(SubProblemBase):
         if self.solver == "cbc":
             self.prob.solve(pulp.PULP_CBC_CMD(msg=False, timeLimit=time_limit))
         elif self.solver == "cplex":
-            self.prob.solve(pulp.CPLEX_CMD(msg=False, timelimit=time_limit))
+            self.prob.solve(pulp.CPLEX_CMD(msg=False, timeLimit=time_limit))
         elif self.solver == "gurobi":
             gurobi_options = []
             if time_limit is not None:
-                gurobi_options.append((
-                    "TimeLimit",
-                    time_limit,
-                ))
+                gurobi_options.append(("TimeLimit", time_limit,))
             self.prob.solve(pulp.GUROBI(msg=False, options=gurobi_options))
 
     def _formulate(self):
         # minimize reduced cost
-        self.prob += pulp.lpSum([
-            self.sub_G.edges[i, j]["weight"] * self.x[(i, j)]
-            for (i, j) in self.sub_G.edges()
-        ])
+        self.prob += pulp.lpSum(
+            [
+                self.sub_G.edges[i, j]["weight"] * self.x[(i, j)]
+                for (i, j) in self.sub_G.edges()
+            ]
+        )
         # flow balance
         for v in self.sub_G.nodes():
             if v not in ["Source", "Sink"]:
                 in_flow = pulp.lpSum(
-                    [self.x[(i, v)] for i in self.sub_G.predecessors(v)])
+                    [self.x[(i, v)] for i in self.sub_G.predecessors(v)]
+                )
                 out_flow = pulp.lpSum(
[…]

[…]
