### 2020-09-03 Kuifje02/vrpy#44 @Kuifje02
I think the revised paper is satisfactory, although maybe too long for JOSS requirements. I am closing for now.
