### 2020-09-04 ff66329
replaced np.matrix by np.array and asserted results #53
