### 2020-09-05 #2408 @skadio
I do not have an issue with the slightly longer resulting paper --as it is adding value. Same for the additional author.

That said, I cannot comment on these "editorially" from JOSS perspective but only as a reviewer.
