### 2020-09-05 #2408 @arfon
> @kakiac @bstabler While we are at it, another question. We have taken into account all of the comments for the paper, but the resulting paper is longer than the 1000 word limit (~ 1300 words). Is the 1000 word limit strict ?

It's OK to go a little over so as long as @kakiac is OK with this you should be fine.
