### 2020-09-08 0938587
replaced numpy matrix by numpy array #53
