### 2020-09-13 #2408 @skadio
Let me follow-up on the status submission; what are the next steps to make this work to cross the finish line? The work & review seem almost complete but I am not sure how to finalize this.

@Kuifje02 my last question about the submission is about the "Community Guidelines". Please see the open review item and let us know your thoughts.

@bstabler I see that you completed most of the review checklist, Any following comments/questions from your end?
