### 2020-09-14 #2408 @Kuifje02
Thanks @skadio for the follow up.

The Community Guidelines have been addressed with a [CONTRIBUTING.md](https://github.com/Kuifje02/vrpy/blob/dev/CONTRIBUTING.md) file, which is referenced at the bottom of the [README.md](https://github.com/Kuifje02/vrpy/tree/dev).

We still have a few minor items to do :

- [ ]  rename main.py to vehicle_routing.py

- [ ] add a license file

- [ ] complete the public API docs (?)

Hopefully we can do this quickly (the two first items for sure), merge with the master branch, and the finish line should not be so far :)
