### 2020-09-14 Kuifje02/vrpy#53 @Kuifje02
I removed the examples/notebook, as everything is in the docs and the tests.
