### 2020-09-14 0386831
minimize global span option
