+++ vrpy/master_solve_pulp.py
@@ -28,18 +28,22 @@ class MasterSolvePulp(MasterProblemBase):
         self.drop = {}  # dropping variable
         self.dummy = {}  # dummy variable
         self.dummy_bound = {}  # dummy variable for vehicle bound cost
+        self.makespan = {}  # maximum global span
         # constrs
         self.set_covering_constrs = {}
         self.vehicle_bound_constrs = {}
         self.drop_penalty_constrs = {}
+        self.makespan_constr = {}
         # Restricted master heuristic
         self.diving_heuristic = DivingHeuristic()
+        # Parameter when minimizing global span
+        self._n_columns = 1000
 
         self._formulate()
 
     def solve(self, relax, time_limit):
         self._solve(relax, time_limit)
-        logger.debug("master problem")
+        logger.debug("master problem relax %s" % relax)
         logger.debug("Status: %s" % pulp.LpStatus[self.prob.status])
         logger.debug("Objective: %s" % pulp.value(self.prob.objective))
 
@@ -47,11 +51,11 @@ class MasterSolvePulp(MasterProblemBase):
             raise Exception("problem " + str(pulp.LpStatus[self.prob.status]))
         if relax:
             for r in self.routes:
-                if pulp.value(self.y[r.graph["name"]]) > 0.5:
-                    logger.debug("route %s selected" % r.graph["name"])
+                val = pulp.value(self.y[r.graph["name"]])
+                if val > 0.1:
+                    logger.debug("route %s selected %s" % (r.graph["name"], val))
         duals = self.get_duals()
         logger.debug("duals : %s" % duals)
-
         return duals, self.prob.objective.value()
 
     def solve_and_dive(self, time_limit):
@@ -63,7 +67,10 @@ class MasterSolvePulp(MasterProblemBase):
     def update(self, new_route):
         """Add new column.
         """
-        self._add_route_selection_variable(new_route)
+        if self.minimize_global_span:
+            self._add_route_selection_variable_for_global_span(new_route)
+        else:
+            self._add_route_selection_variable(new_route)
 
     def get_duals(self, relax: pulp.LpProblem = None):
         """Gets the dual values of each constraint of the master problem.
@@ -97,6 +104,17 @@ class MasterSolvePulp(MasterProblemBase):
                 else:
                     duals["upper_bound_vehicles"][k] = relax.constraints[
                         "upper_bound_vehicles_%s" % k
+                    ].pi
+        # global span duals
+        if self.minimize_global_span:
+            for route in self.routes:
+                if not relax:
+                    duals["makespan_%s" % route.graph["name"]] = self.prob.constraints[
+                        "makespan_%s" % route.graph["name"]
+                    ].pi
+                else:
+                    duals["makespan_%s" % route.graph["name"]] = relax.constraints[
+                        "makespan_%s" % route.graph["name"]
                     ].pi
         return duals
 
@@ -159,6 +177,7 @@ class MasterSolvePulp(MasterProblemBase):
                 if "artificial_bound_" in var.name:
                     var.upBound = 0
                     var.lowBound = 0
+        self.prob.writeLP("master.lp")
         # Solve with appropriate solver
         if self.solver == "cbc":
             self.prob.solve(
@@ -199,10 +218,20 @@ class MasterSolvePulp(MasterProblemBase):
         if self.num_vehicles and not self.periodic:
             self._add_bound_vehicles()
 
+        if self.minimize_global_span:
+            self._add_maximum_makespan_constraints()
+
         # Add variables #
         # Route selection variables
-        for route in self.routes:
-            self._add_route_selection_variable(route)
+        if self.minimize_global_span:
[…]
