+++ vrpy/__init__.py
@@ -1,2 +1,2 @@
 """vrpy modules."""
-from vrpy.main import VehicleRoutingProblem
+from vrpy.vrp import VehicleRoutingProblem
