### 2020-09-14 Kuifje02/vrpy#43 @Kuifje02
@torressa What do you think about this one ?

> * Is vrpy.VehicleRoutingProblem the only "public" interface? I believe so, but if not, please add others to API.html. Btw, maybe add all other classes the prefix "_" to hint the private nature?

VehicleRoutingProblem is our only public class. Is it worth redefining all other classes with a prefix (do you know if this is a standard norm) ?

And do you think the API should be more documented ? I'm afraid it might be redundant with all the examples.
