### 2020-09-15 Kuifje02/vrpy#43 @torressa
The private class thing is a fair point as I think it's pretty standard: see [here](https://docs.python.org/3/tutorial/classes.html#private-variables).

I'm with you on the API (which rtd doesn't seem to like now) so maybe adding a link to the examples page would make the reviewers happy
