+++ examples/vrptw.py
@@ -10,7 +10,7 @@ from numpy import array
 from data import DISTANCES, TRAVEL_TIMES, TIME_WINDOWS_LOWER, TIME_WINDOWS_UPPER
 import sys
 
-sys.path.append("../../")
+sys.path.append("../")
 from vrpy import VehicleRoutingProblem
 
 # Transform distance matrix to DiGraph
