+++ examples/cvrp.py
@@ -3,7 +3,7 @@ from numpy import array
 from data import DISTANCES, DEMANDS
 import sys
 
-sys.path.append("../../")
+sys.path.append("../")
 from vrpy import VehicleRoutingProblem
 
 # Transform distance matrix to DiGraph
@@ -19,7 +19,7 @@ G = relabel_nodes(G, {0: "Source", 17: "Sink"})
 if __name__ == "__main__":
 
     prob = VehicleRoutingProblem(G, load_capacity=15)
-    prob.solve(solver="cplex", cspy=False, time_limit=0)
+    prob.solve(pricing_strategy="Hyper")
     print(prob.best_value)
     print(prob.best_routes)
     assert prob.best_value == 6208
