### 2020-09-15 89ea314
if heterogeneous fleet, go back to master after first -ve column found
