+++ vrpy/clarke_wright.py
@@ -1,7 +1,7 @@
 from networkx import DiGraph, add_path, shortest_path
 
 
-class ClarkeWright:
+class _ClarkeWright:
     """
     Clarke & Wrights savings algorithm.
 
@@ -11,6 +11,7 @@ class ClarkeWright:
         duration (int, optional) : Maximum duration per route. Defaults to None.
         num_stops (int, optional) : Maximum number of stops per route. Defaults to None.
     """
+
     def __init__(
         self,
         G,
@@ -61,8 +62,9 @@ class ClarkeWright:
         for v in self.G.nodes():
             if v not in ["Source", "Sink"]:
                 # Create round trip
-                round_trip_cost = (self.G.edges["Source", v]["cost"] +
-                                   self.G.edges[v, "Sink"]["cost"])
+                round_trip_cost = (
+                    self.G.edges["Source", v]["cost"] + self.G.edges[v, "Sink"]["cost"]
+                )
                 route = DiGraph(cost=round_trip_cost)
                 add_path(route, ["Source", v, "Sink"])
                 self._route[v] = route
@@ -70,9 +72,11 @@ class ClarkeWright:
                 if self.load_capacity:
                     route.graph["load"] = self.G.nodes[v]["demand"]
                 if self.duration:
-                    route.graph["time"] = (self.G.nodes[v]["service_time"] +
-                                           self.G.edges["Source", v]["time"] +
-                                           self.G.edges[v, "Sink"]["time"])
+                    route.graph["time"] = (
+                        self.G.nodes[v]["service_time"]
+                        + self.G.edges["Source", v]["time"]
+                        + self.G.edges[v, "Sink"]["time"]
+                    )
 
     def _update_routes(self):
         """Stores best routes as list of nodes."""
@@ -86,9 +90,9 @@ class ClarkeWright:
         for (i, j) in self.G.edges():
             if i != "Source" and j != "Sink":
                 self._savings[(i, j)] = (
-                    self.G.edges[i, "Sink"]["cost"] +
-                    self.G.edges["Source", j]["cost"] -
-                    self.alpha * self.G.edges[i, j]["cost"]
+                    self.G.edges[i, "Sink"]["cost"]
+                    + self.G.edges["Source", j]["cost"]
+                    - self.alpha * self.G.edges[i, j]["cost"]
                     # FOR MORE SOPHISTICATED VERSIONS OF CLARKE WRIGHT:
                     # + self.beta
                     # * abs(
@@ -99,9 +103,7 @@ class ClarkeWright:
                     # * (self.G.nodes[i]["demand"] + self.G.nodes[j]["demand"])
                     # / self._average_demand
                 )
-        self._ordered_edges = sorted(self._savings,
-                                     key=self._savings.get,
-                                     reverse=True)
+        self._ordered_edges = sorted(self._savings, key=self._savings.get, reverse=True)
 
     def _merge_route(self, existing_node, new_node, depot):
         """
@@ -119,9 +121,10 @@ class ClarkeWright:
             route.remove_edge(existing_node, "Sink")
             # Update route cost
             self._route[existing_node].graph["cost"] += (
-                self.G.edges[existing_node, new_node]["cost"] +
-                self.G.edges[new_node, "Sink"]["cost"] -
-                self.G.edges[existing_node, "Sink"]["cost"])
+                self.G.edges[existing_node, new_node]["cost"]
+                + self.G.edges[new_node, "Sink"]["cost"]
+                - self.G.edges[existing_node, "Sink"]["cost"]
+            )
 
         # Insert new_node between Source and existing_node
         if depot == "Source":
@@ -129,21 +132,22 @@ class ClarkeWright:
             route.remove_edge("Source", existing_node)
             # Update route cost
             self._route[existing_node].graph["cost"] += (
[…]
