+++ vrpy/subproblem.py
@@ -13,7 +13,7 @@ from networkx import (
 logger = logging.getLogger(__name__)
 
 
-class SubProblemBase:
+class _SubProblemBase:
     """
     Base class for the subproblems.
 
