+++ docs/refs.bib
@@ -1,25 +1,3 @@
-@Misc{johnjforrest_2020,
-  author       = {johnjforrest and
-                  Stefan Vigerske and
-                  Ted Ralphs and
-                  Lou Hafer and
-                  jpfasano and
-                  Haroldo Gambini Santos and
-                  Matthew Saltzman and
-                  h-i-gassmann and
-                  Bjarni Kristjansson and
-                  Alan King},
-  title        = {coin-or/Clp: Version 1.17.6},
-  month        = apr,
-  year         = 2020,
-  publisher    = {Zenodo},
-  version      = {releases/1.17.6},
-  doi          = {10.5281/zenodo.3748677},
-  url          = {https://doi.org/10.5281/zenodo.3748677}
-}
-
-
-
 
 @Misc{cspy,
   author = {Torres Sanchez, David},
@@ -114,6 +92,31 @@
   publisher={Elsevier}
 }
 
+@inproceedings{sabar2015math,
+  title={A math-hyper-heuristic approach for large-scale vehicle routing problems with time windows},
+  author={Sabar, Nasser R and Zhang, Xiuzhen Jenny and Song, Andy},
+  booktitle={2015 IEEE Congress on Evolutionary Computation (CEC)},
+  pages={830--837},
+  year={2015},
+  organization={IEEE}
+}
 
+@inproceedings{ferreira2017multi,
+  title={A multi-armed bandit selection strategy for hyper-heuristics},
+  author={Ferreira, Alexandre Silvestre and Gon{\c{c}}alves, Richard Aderbal and Pozo, Aurora},
+  booktitle={2017 IEEE Congress on Evolutionary Computation (CEC)},
+  pages={525--532},
+  year={2017},
+  organization={IEEE}
+}
+
+@inproceedings{drake2012improved,
+  title={An improved choice function heuristic selection for cross domain heuristic search},
+  author={Drake, John H and {\"O}zcan, Ender and Burke, Edmund K},
+  booktitle={International Conference on Parallel Problem Solving from Nature},
+  pages={307--316},
+  year={2012},
+  organization={Springer}
+}
 
 
