+++ docs/solving_options.rst
@@ -1,3 +1,5 @@
+.. _options:
+
 Solving Options
 ===============
 
