### 2020-09-15 42527de
solve with default parameters
