### 2020-09-24 c375908
Clean up master
- add Halvard's ORCID no
- add examples/README
- remove print statements
