+++ examples/cvrp_drop.py
@@ -1,9 +1,8 @@
 from networkx import from_numpy_matrix, set_node_attributes, relabel_nodes, DiGraph
 from numpy import array
-from data import DISTANCES, DEMANDS_DROP
-import sys
 
-sys.path.append("../")
+from examples.data import DISTANCES, DEMANDS_DROP
+
 from vrpy import VehicleRoutingProblem
 
 # Transform distance matrix to DiGraph
@@ -18,7 +17,10 @@ G = relabel_nodes(G, {0: "Source", 17: "Sink"})
 
 if __name__ == "__main__":
 
-    prob = VehicleRoutingProblem(G, load_capacity=15, drop_penalty=1000, num_vehicles=4)
+    prob = VehicleRoutingProblem(G,
+                                 load_capacity=15,
+                                 drop_penalty=1000,
+                                 num_vehicles=4)
     prob.solve()
     print(prob.best_value)
     print(prob.best_routes)
