+++ examples/pdp.py
@@ -1,9 +1,8 @@
 from networkx import from_numpy_matrix, relabel_nodes, DiGraph
 from numpy import array
-from data import DISTANCES, PICKUPS_DELIVERIES
-import sys
 
-sys.path.append("../")
+from examples.data import DISTANCES, PICKUPS_DELIVERIES
+
 from vrpy import VehicleRoutingProblem
 
 # Transform distance matrix to DiGraph
@@ -21,7 +20,10 @@ G = relabel_nodes(G, {0: "Source", 17: "Sink"})
 
 if __name__ == "__main__":
 
-    prob = VehicleRoutingProblem(G, load_capacity=6, pickup_delivery=True, num_stops=6)
+    prob = VehicleRoutingProblem(G,
+                                 load_capacity=6,
+                                 pickup_delivery=True,
+                                 num_stops=6)
     prob.solve(cspy=False)
     print(prob.best_value)
     print(prob.best_routes)
