+++ examples/README.md
@@ -1,0 +1,22 @@
+# Examples
+
+- [`cvrp.py`](examples/cvrp.py) - Capacitated vehicle routing
+- [`cvrp_drop.py`](examples/cvrp_drop.py) - Capacitated vehicle routing with dropping penalty
+- [`cvrpsdc.py`](examples/cvrpsdc.py) - Capacitated vehicle routing with distribution and collection
+- [`pdp.py`](examples/pdp.py) - Capacitated vehicle routing with pickup and delivery
+- [`vrptw.py`](examples/vrptw.py) - Vehicle routing with time windows
+
+## Run
+
+To run an example, from the main repo folder do
+
+```bash
+python3 -m examples.<name>
+```
+
+Where `<name>` is the file name (without the extension) above.
+For example
+
+```bash
+python3 -m examples.cvrp
+```
