+++ examples/README.md
@@ -1,10 +1,10 @@
 # Examples
 
-- [`cvrp.py`](examples/cvrp.py) - Capacitated vehicle routing
-- [`cvrp_drop.py`](examples/cvrp_drop.py) - Capacitated vehicle routing with dropping penalty
-- [`cvrpsdc.py`](examples/cvrpsdc.py) - Capacitated vehicle routing with distribution and collection
-- [`pdp.py`](examples/pdp.py) - Capacitated vehicle routing with pickup and delivery
-- [`vrptw.py`](examples/vrptw.py) - Vehicle routing with time windows
+- [`cvrp.py`](cvrp.py) - Capacitated vehicle routing
+- [`cvrp_drop.py`](cvrp_drop.py) - Capacitated vehicle routing with dropping penalty
+- [`cvrpsdc.py`](cvrpsdc.py) - Capacitated vehicle routing with distribution and collection
+- [`pdp.py`](pdp.py) - Capacitated vehicle routing with pickup and delivery
+- [`vrptw.py`](vrptw.py) - Vehicle routing with time windows
 
 ## Run
 
