### 2020-09-29 8cc0c05
Add run_exact parameter to vrp.solve
- Partial merge with master
- Fix runner (actually stops at 1 iteration)
