+++ vrpy/hyper_heuristic.py
@@ -1,11 +1,11 @@
-from random import uniform, choice
 from math import exp, log, sqrt
 from time import time
 import logging
 import bisect
 
+from vrpy.checks import check_seed
+
 logger = logging.getLogger(__name__)
-logging.basicConfig(level=logging.INFO)
 
 
 class _HyperHeuristic:
@@ -28,13 +28,19 @@ class _HyperHeuristic:
             heuristic_options: "weighted_average", "relative_improvement"
             Defaults to "weighted_average".
         acceptance_type(str):
-            Set move acceptance. Defaults to "Accept all"
+            Set move acceptance.
+            - `accept_all`
+            - `table` - measure
+            - `ob
+            Defaults to "accept_all"
         step (float, optional):
             Adaptive step.
             Defaults to 0.02.
         start_computing_average (int):
             Iteration where average starts being computed.
             Defaults to 1.
+        seed (int or numpy.random.RandomState):
+            Defaults to None.
     """
 
     def __init__(
@@ -42,9 +48,10 @@ class _HyperHeuristic:
         heuristic_options=["BestPaths", "BestEdges1", "BestEdges2", "Exact"],
         scaling_factor: float = 0.5,
         performance_measure_type="weighted_average",
-        acceptance_type="Accept all",
+        acceptance_type="objective_decrease",
         step=0.1,
         start_computing_average=1,
+        seed=None,
     ):
         # Input parameters
         self.heuristic_options = heuristic_options
@@ -53,6 +60,7 @@ class _HyperHeuristic:
         self.acceptance_type = acceptance_type
         self.step = step
         self.start_computing_average = start_computing_average
+        self.random_state = check_seed(seed)
 
         # Internal parameters
         self.theta = 1
@@ -85,10 +93,10 @@ class _HyperHeuristic:
         # Weights
         # See _update_params_wgtavr for use
         self.weight_col_basic = 0.5
-        self.weight_runtime = 0.5
+        self.weight_runtime = 0.1
         self.weight_spread = 0.3
-        self.weight_obj = 0.2
-        self.weight_col_total = 0.1
+        self.weight_obj = 0.05
+        self.weight_col_total = 0.05
 
         # time related parameters
         self.start_time = None
@@ -112,7 +120,7 @@ class _HyperHeuristic:
         if self.performance_measure_type == "relative_improvement":
             # before the number of iterations is high enough
             if self.iteration < self.start_computing_average:
-                return choice(self.heuristic_options[0:2])
+                return self.random_state.choice(self.heuristic_options[0:2])
         elif self.performance_measure_type == "weighted_average":
             pass
         # choose according to MAB
@@ -123,7 +131,7 @@ class _HyperHeuristic:
         if len(best_heuristics) == 1:
             self.current_heuristic = best_heuristics[0]
         else:
-            self.current_heuristic = choice(best_heuristics)
+            self.current_heuristic = self.random_state.choice(best_heuristics)
         return self.current_heuristic
 
     def update_scaling_factor(self, no_improvement_count: int,
@@ -172,8 +180,8 @@ class _HyperHeuristic:
 
         elif self.performance_measure_type == "weighted_average":
             self._current_performance_wgtavr(
-                active_columns=active_columns, new_objective_value=new_objective_value
-            )
+                active_columns=active_columns,
+                new_objective_value=new_objective_value)
         else:
             raise ValueError("performence_measure not set correctly!")
         self.set_current_objective(objective=new_objective_value)
@@ -187,15 +195,15 @@ class _HyperHeuristic:
             new_objective_value (float): new_objective_value. Defaults to None.
         """
         update = True
-        if self.acceptance_type == "Accept all":
+        if self.acceptance_type == "accept_all":
             return update
-        elif self.acceptance_type == "Table":
+        elif self.acceptance_type == "table":
[…]
