+++ vrpy/old/master_solve_pulp.py
@@ -1,353 +1,0 @@
-import pulp
-import logging
-from networkx import shortest_path
-
-from vrpy.masterproblem import MasterProblemBase
-
-logger = logging.getLogger(__name__)
-
-
-class MasterSolvePulp(MasterProblemBase):
-    """
-    Solves the master problem for the column generation procedure.
-
-    Inherits problem parameters from MasterProblemBase
-    """
-
-    def solve(self):
-        self._formulate()
-        self.prob.writeLP("master.lp")
-        self._solve()
-        logger.debug("master problem")
-        logger.debug("Status: %s" % pulp.LpStatus[self.prob.status])
-        logger.debug("Objective: %s" % pulp.value(self.prob.objective))
-
-        if pulp.LpStatus[self.prob.status] != "Optimal":
-            raise Exception("problem " + str(pulp.LpStatus[self.prob.status]))
-        if self.relax:
-            for r in self.routes:
-                if pulp.value(self.y[r.graph["name"]]) > 0.5:
-                    logger.debug("route %s selected" % r.graph["name"])
-            duals = self.get_duals()
-            logger.debug("duals : %s" % duals)
-            return duals, pulp.value(self.prob.objective)
-
-        else:
-            return self._get_total_cost_and_routes()
-
-    def solve_and_dive(self, max_depth=3, max_discrepancy=1):
-        """
-        Implements diving algorithm with Limited Discrepancy Search
-        Parameters as suggested by the authors. This only fixes one column.
-        `Sadykov et al. (2019)`_.
-
-        .. _Sadykov et al. (2019): https://pubsonline.informs.org/doi/abs/10.1287/ijoc.2018.0822
-        """
-        self._formulate()
-        self._solve()
-
-        depth = 0
-        tabu_list = []
-        stop_diving = True
-        relax = self.relax.model.deepcopy()
-        constrs = {}
-        while depth <= max_depth and len(tabu_list) <= max_discrepancy:
-            non_integer_vars = list(
-                var
-                for var in relax.variables()
-                if abs(var.varValue - round(var.varValue)) != 0
-            )
-            # All non-integer variables not already fixed in this or any
-            # iteration of the diving heuristic
-            vars_to_fix = [
-                var
-                for var in non_integer_vars
-                if var.name not in self._tabu_list and var.name not in tabu_list
-            ]
-            if vars_to_fix:
-                # If non-integer variables not already fixed and
-                # max_discrepancy not violated
-                stop_diving = False
-
-                var_to_fix = min(
-                    vars_to_fix, key=lambda x: abs(x.varValue - round(x.varValue))
-                )
-                value_to_fix = 1
-                value_previous = var_to_fix.varValue
-
-                name_le = "fix_{}_LE".format(var_to_fix.name)
-                name_ge = "fix_{}_GE".format(var_to_fix.name)
-                constrs[name_le] = pulp.LpConstraint(
-                    var_to_fix, pulp.LpConstraintLE, name=name_le, rhs=value_to_fix
-                )
-                constrs[name_ge] = pulp.LpConstraint(
-                    var_to_fix, pulp.LpConstraintGE, name=name_ge, rhs=value_to_fix
-                )
-
-                relax += constrs[name_le]  # add <= constraint
-                relax += constrs[name_ge]  # add >= constraint
-                relax.solve()
-                # if not optimal status code from :
-                # https://github.com/coin-or/pulp/blob/master/pulp/constants.py#L45-L57
-                if relax.status != 1 or all(
-                    abs(v.varValue - round(v.varValue)) == 0 for v in relax.variables()
-                ):
-                    stop_diving = True
-                    break
-                if len(tabu_list) >= max_discrepancy:
-                    break
-                tabu_list.append(var_to_fix.name)
-                depth += 1
[…]
