+++ vrpy/vrp.py
@@ -119,14 +119,13 @@ class VehicleRoutingProblem:
         self._dive: bool = None
         self._greedy: bool = None
         self._max_iter: int = None
-        self._return_runtime: bool = None
+        self._run_exact = None  # iterations after which the exact algorithm is ran
 
         # parameters for column generation stopping criteria
         self._start_time = None
         self._more_routes = None
         self._iteration = 0  # current iteration
         self._no_improvement = 0  # iterations after with no change in obj func
-        self._run_exact = 10  # iterations after which the exact algorithm is ran
         self._lower_bound = []
         # Parameters for initial solution and preprocessing
         self._max_capacity: int = None
@@ -154,6 +153,7 @@ class VehicleRoutingProblem:
         dive=False,
         greedy=False,
         max_iter=None,
+        run_exact=1,
     ):
         """Iteratively generates columns with negative reduced cost and solves as MIP.
 
@@ -201,9 +201,10 @@ class VehicleRoutingProblem:
                 Defaults to False.
             max_iter (int, optional):
                 maximum number of iterations for the column generation procedure.
-            return_runtime (bool, optional):
-                True if solve runtime is to be returned
-                Defaults to False.
+            run_exact (int, optional):
+                if a pricing strategy is selected, this parameter controls the number of
+                iterations after which the exact algorithm is run.
+                Defaults to 1.
 
         Returns:
             float: Optimal solution of MIP based on generated columns
@@ -218,6 +219,7 @@ class VehicleRoutingProblem:
         self._dive = False
         self._greedy = greedy
         self._max_iter = max_iter
+        self._run_exact = run_exact
         if self._pricing_strategy == "Hyper":
             self.hyper_heuristic = _HyperHeuristic()
 
@@ -263,25 +265,23 @@ class VehicleRoutingProblem:
         """Returns dict with route ids as keys and route costs as values."""
         cost = {}
         for route in self.best_routes:
-            edges = list(zip(self.best_routes[route][:-1], self.best_routes[route][1:]))
+            edges = list(
+                zip(self.best_routes[route][:-1], self.best_routes[route][1:]))
             k = self._best_routes_vehicle_type[route]
-            cost[route] = sum(self._H.edges[i, j]["cost"][k] for (i, j) in edges)
+            cost[route] = sum(
+                self._H.edges[i, j]["cost"][k] for (i, j) in edges)
         return cost
 
     @property
     def best_routes_load(self):
         """Returns dict with route ids as keys and route loads as values."""
         load = {}
-        if (
-            not self.load_capacity
-            or self.distribution_collection
-            or self.pickup_delivery
-        ):
+        if (not self.load_capacity or self.distribution_collection or
+                self.pickup_delivery):
             return load
         for route in self.best_routes:
             load[route] = sum(
-                self._H.nodes[v]["demand"] for v in self.best_routes[route]
-            )
+                self._H.nodes[v]["demand"] for v in self.best_routes[route])
         return load
 
     @property
@@ -293,11 +293,8 @@ class VehicleRoutingProblem:
         If truck is distributing, load refers to accumulated amount that has been unloaded.
         """
         load = {}
-        if (
-            not self.load_capacity
-            and not self.pickup_delivery
-            and not self.distribution_collection
-        ):
+        if (not self.load_capacity and not self.pickup_delivery and
+                not self.distribution_collection):
             return load
         for i in self.best_routes:
             load[i] = {}
@@ -317,13 +314,14 @@ class VehicleRoutingProblem:
         if not self.duration and not self.time_windows:
             return duration
         for route in self.best_routes:
-            edges = list(zip(self.best_routes[route][:-1], self.best_routes[route][1:]))
+            edges = list(
+                zip(self.best_routes[route][:-1], self.best_routes[route][1:]))
             # Travel times
[…]
