### 2020-09-30 #2408 @kakiac
:wave @Kuifje02 @skadio @bstabler @arfon many thanks for all your contributions and great work on the paper! Just going through the final checks for the paper to finalise and publish 🎆

* in reply to https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-671821850 - yes, fine to have an additional author included, I can see his contribution evidenced as well on the repository. @Halvaros is that ok with you?
* in reply to https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-671830216 - very happy with the additional length, as it adds value to the submission (and it is not excessive!)

Here are a few last bits:

@Kuifje02
[  ] could you merge my suggested minor change/typo on your paper.bib: https://github.com/Kuifje02/vrpy/pull/64

@skadio
* can you confirm (by ticking this box or the box in your review checklist) that you are happy that the following have been adequately addressed in the revised version of the paper:

  [  ]  Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

* would you recommend acceptance?

@bstabler
* can you confirm (by ticking this box or the box in your review checklist) that you are happy that the following have been adequately addressed in the revised version of the paper:
  [  ] **Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
  [  ] **Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
  [  ] **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
  [  ]  **State of the field**: Do the authors describe how this software compares to other commonly-used packages?

* would you recommend acceptance?
