### 2020-09-30 16a5c9f
with added dois, where missing, on bibliography
I have added dois for the papers here - I could find all referenced here except for reference: @cordeau2000vrp. Can you please check my proposed changes and add the url or doi for that paper as well? I can then run the final checks for https://github.com/openjournals/joss-reviews/issues/2408
