+++ paper/paper.bib
@@ -18,7 +18,8 @@
   number={4},
   pages={946--985},
   year={2019},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi = {10.1287/trsc.2018.0878}
 }
 
 @article{cspy,
@@ -38,7 +39,8 @@
   title={Exploring network structure, dynamics, and function using NetworkX},
   author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
   year={2008},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)},
+  url = {https://conference.scipy.org/proceedings/scipy2008/paper_2/full_text.pdf}
 }
 
 @Misc{numpy,
@@ -58,7 +60,8 @@
   number={2},
   pages={235--247},
   year={2006},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi = {10.1287/trsc.1050.0118}
 }
 
 @Misc{ortools,
@@ -78,7 +81,8 @@
   number={4},
   pages={568--581},
   year={1964},
-  publisher={Informs}
+  publisher={INFORMS},
+  doi = {10.1287/opre.12.4.568}
 }
 
 @software{johnjforrest_2020_Cbc,
@@ -130,7 +134,8 @@
   booktitle={The Logic of Logistics},
   pages={125--141},
   year={1997},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/0-387-22619-2_16}
 }
 
 @article{dantzig1959truck,
@@ -141,7 +146,8 @@
   number={1},
   pages={80--91},
   year={1959},
-  publisher={Informs}
+  publisher={INFORMS},
+  doi = {10.1287/mnsc.6.1.80}
 }
 
 @incollection{desrosiers1988shortest,
@@ -150,7 +156,8 @@
   booktitle={Advances in Optimization and Control},
   pages={144--157},
   year={1988},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-642-46629-8_10}
 }
 
 @article{choi2007column,
@@ -161,7 +168,8 @@
   number={7},
   pages={2080--2095},
   year={2007},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1016/j.cor.2005.08.002}
 }
 
 @book{cordeau2000vrp,
@@ -179,7 +187,8 @@
   number={5},
   pages={1050--1073},
   year={1985},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi={10.1287/opre.33.5.1050}
 }
 
 @article{baldacci2010exact,
@@ -190,7 +199,8 @@
   number={1},
   pages={213--245},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/s10479-009-0650-0}
 }
 
 @article{laporte2007you,
@@ -201,7 +211,8 @@
   number={8},
   pages={811--819},
   year={2007},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1002/nav.20261}
 }
 
 @article{sadykov2019primal,
@@ -212,7 +223,8 @@
   number={2},
   pages={251--267},
   year={2019},
-  publisher={INFORMS}
+  publisher={INFORMS},
+  doi = {10.1287/ijoc.2018.0822}
 }
 
 @article{solomon1987algorithms,
@@ -223,7 +235,8 @@
   number={2},
   pages={254--265},
   year={1987},
-  publisher={Informs}
+  publisher={INFORMS},
[…]
