### 2020-09-30 #2408 @skadio
@kakiac  @Kuifje02 Yes, I confirm the changes which improve the library/paper. I assume that the author can easily fix the remaining missing bib, etc. With that, my recommendation is to accept.
