### 2020-10-01 #2408 @Kuifje02
> * in reply to [#2408 (comment)](https://github.com/openjournals/joss-reviews/issues/2408#issuecomment-671821850) - yes, fine to have an additional author included, I can see his contribution evidenced as well on the repository. @Halvaros is that ok with you?

This is ok with @Halvaros 👍

> Here are a few last bits:
>
> @Kuifje02
>
> * [x]  could you merge my suggested minor change/typo on your paper.bib: [Kuifje02/vrpy#64](https://github.com/Kuifje02/vrpy/pull/64)
> * [x]  check / accept changes on DOIs in paper.bib for the references in this issue: [Kuifje02/vrpy#65](https://github.com/Kuifje02/vrpy/pull/65)
> * [x]  minor typo :) [Kuifje02/vrpy#66](https://github.com/Kuifje02/vrpy/pull/66)
> * [x]  references ok
> * [x]  typos, wording etc ok
