+++ paper/paper.bib
@@ -177,6 +177,7 @@
   author={Cordeau, Jean-Francois and Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions (Montr{\'e}al, Qu{\'e}bec)},
   year={2000},
   publisher={Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions Montr{\'e}al}
+  url={https://doi.org/10.1137/1.9780898718515.ch7}
 }
 
 @article{laporte1985optimal,
