+++ paper/paper.bib
@@ -177,7 +177,7 @@
   author={Cordeau, Jean-Francois and Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions (Montr{\'e}al, Qu{\'e}bec)},
   year={2000},
   publisher={Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions Montr{\'e}al}
-  url={https://doi.org/10.1137/1.9780898718515.ch7}
+  url={https://pdfs.semanticscholar.org/3aaa/a16ab53cf30c378fdb7c911fe0de39ee8997.pdf}
 }
 
 @article{laporte1985optimal,
