+++ paper/paper.bib
@@ -176,7 +176,7 @@
   title={The VRP with time windows},
   author={Cordeau, Jean-Francois and Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions (Montr{\'e}al, Qu{\'e}bec)},
   year={2000},
-  publisher={Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions Montr{\'e}al}
+  publisher={Groupe d'{\'e}tudes et de recherche en analyse des d{\'e}cisions Montr{\'e}al},
   url={https://pdfs.semanticscholar.org/3aaa/a16ab53cf30c378fdb7c911fe0de39ee8997.pdf}
 }
 
