### 2020-10-15 #2408 @arfon
Him, Crossref's API is struggling with some of these references. Not sure why...
