### 2020-10-15 #2408 @arfon
Fixed this in https://github.com/openjournals/whedon-api/commit/8e4d5d1576ca4bfc1a25adca151f281496501283
