### 2020-11-05 #2408 @kakiac
> 👋 @kakiac - how are we doing here? Is this paper nearly ready to go?

Yes, last few checks, not too long to go now!
