### 2020-11-05 #2408 @kakiac
👋 @bstabler would you recommend acceptance for this submission to JOSS? Thank you!
