### 2020-11-05 #2408 @arfon
> 👋 @bstabler would you recommend acceptance for this submission to JOSS? I think you have checked the checklist, but I would be grateful if you could confirm in the thread as well for the record 😄

@kakiac - I know that @bstabler has very limited availability right now so I wouldn't wait more than ~24 hours for a response here.
