### 2020-11-05 #2408 @kakiac
@arfon Many thanks will keep that in mind 😄

@Kuifje02, in the meantime, could you merge these last minor citation changes:

- https://github.com/Kuifje02/vrpy/pull/68
- https://github.com/Kuifje02/vrpy/pull/67

After that, can you please:
 - make a tagged release (on github) and
 - archive (on Zenodo or Figshare), and
 - report the tagged version number and archive DOI in this review thread.
