### 2020-11-06 #2408 @Kuifje02
> After that, can you please:
>
> * make a tagged release (on github) and

https://github.com/Kuifje02/vrpy/releases/tag/0.3.0

> * archive (on Zenodo or Figshare), and

https://zenodo.org/record/4248877

> * report the tagged version number and archive DOI in this review thread.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4248877.svg)](https://doi.org/10.5281/zenodo.4248877)
