### 2020-11-09 #2408 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
