### 2020-11-09 #2408 @kthyng
Congratulations on your new publication @Kuifje02! Thanks to editor @kakiac and reviewers @bstabler and @skadio for your time and expertise — we couldn't do this without you!!
