### 2020-11-09 #2408 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
