### 2020-11-10 #2408 @arfon
> Congratulations @Kuifje02! 🥳 🎈 🎉

...and congratulations to @kakiac on the successful completion of their first JOSS paper as editor 🎉
