### 2020-11-10 #2408 @Kuifje02
Thank you everyone !

It was a pleasure working together. Hopefully it was not the last time :) !
